blart::allocator

Module inner

Structs

GlobalExperimental

The global memory allocator.

Traits

AllocatorExperimental

An implementation of Allocator can allocate, grow, shrink, and deallocate arbitrary blocks of data described via Layout.

Functions

do_alloc 🔒