Module blart::bytes

Re-exports

pub use ordered::*;

Modules

ordered 🔒

Macros

as_bytes_for_integer_arrays 🔒

as_bytes_for_integer_like_types 🔒

Traits

AsBytes

Any type implementing AsBytes can be decomposed into bytes.

NoPrefixesBytes

This trait is used to mark types which have a byte representation which is guaranteed to not be a prefix of any other value of the same type.

OrderedBytes

This trait is used to mark types where the lexicographic ordering of their byte representation (as output by AsBytes::as_bytes) matches their normal ordering (as determined by Ord).