blart::bytes::mapped

Struct ToUBE

pub struct ToUBE;

This struct represents a conversion of unsigned integers to the big endian format, so that the natural ordering of the numbers matches the lexicographic ordering of the bytes.