Trait blart::bytes::mapped::BytesMapping

pub trait BytesMapping<D> {
    type Bytes: AsBytes;

    // Required methods
    fn to_bytes(value: D) -> Self::Bytes;
    fn from_bytes(bytes: Self::Bytes) -> D;
}

Trait representing a reversible conversion from a type to some sort of byte string.

Hash and Eq

The transformation from the original type to the byte string should preserve equality for the PartialEq and Eq implementations, along with hashing for the Hash implementation.

The following property should hold true:

value_a == value_b -> Mapping::to_bytes(value_a) == Mapping::to_bytes(value_b)

And this implies that the hash implementations should also match, in line with the Hash and Eq documentation.

Hash and Eq documentation

Ordering

The mapping may optionally preserve ordering of the original type, as implemented by Ord on the original type. In that case, PartialOrd and Ord should be implemented for Mapped<B, D> where B is the type implementing BytesMapping<D> and D is the type being converted. Then the trait OrderedBytes should be implemented for Mapped<B, D> as well.

Required Associated Types

type Bytes: AsBytes

The bytestring type that the D is converted to.

Required Methods

fn to_bytes(value: D) -> Self::Bytes

Convert the domain type into the bytestring type

fn from_bytes(bytes: Self::Bytes) -> D

Convert the bytestring type back into the domain type

Object Safety

This trait is not object safe.