Struct blart::bytes::ordered::ToBE

pub struct ToBE<N>(PhantomData<N>);

This struct represents a conversion of unsigned integers to the big endian format, so that the natural ordering of the numbers matches the lexicographic ordering of the bytes.

Tuple Fields
0: PhantomData<N>