Trait blart::bytes::NoPrefixesBytes

pub unsafe trait NoPrefixesBytes: AsBytes { }

This trait is used to mark types which have a byte representation which is guaranteed to not be a prefix of any other value of the same type.

Safety

This trait can only be implemented if the above condition holds.