Module blart::collections::map::entry_ref

Structs

OccupiedEntryRef

A view into an occupied entry in a TreeMap. It is part of the EntryRef enum.

VacantEntryRef

A view into a vacant entry in a TreeMap. It is part of the EntryRef enum.

Enums

EntryRef

A view into a single entry in a map, which may either be vacant or occupied.