blart::collections::map::iterators

Module prefix

Macros

implement_prefix_iter 🔒

Structs

Prefix

An iterator over a range of entries that all have the same key prefix in a TreeMap.

PrefixMut

A mutable iterator over a range of entries that all have the same key prefix in a TreeMap.

Functions

prefix_iter_constructor 🔒