Struct blart::collections::map::TreeMap

pub struct TreeMap<K, V, const PREFIX_LEN: usize = 16> {
    num_entries: usize,
    pub(crate) root: Option<OpaqueNodePtr<K, V, PREFIX_LEN>>,
}

An ordered map based on an adaptive radix tree.

Fields
num_entries: usize

The number of entries present in the tree.

root: Option<OpaqueNodePtr<K, V, PREFIX_LEN>>

A pointer to the tree root, if present.

Implementations

impl<K, V> TreeMap<K, V>

pub fn new() -> Self

Create a new, empty crate::TreeMap with the default number of prefix bytes (16).

This function will not pre-allocate anything.

Examples

use blart::TreeMap;

let map = TreeMap::<Box<[u8]>, ()>::new();
assert_eq!(map, TreeMap::new());
assert!(map.is_empty());
Run

impl<K, V, const PREFIX_LEN: usize> TreeMap<K, V, PREFIX_LEN>

pub fn with_prefix_len() -> Self

Create a new, empty crate::TreeMap.

This function will not pre-allocate anything.

Examples

use blart::TreeMap;

let map = TreeMap::<Box<[u8]>, (), 16>::with_prefix_len();
assert_eq!(map, TreeMap::new());
assert!(map.is_empty());
Run

pub fn clear(&mut self)

Clear the map, removing all elements.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
assert_eq!(map.len(), 1);

map.clear();
assert!(map.is_empty());
assert!(map.get([1, 2, 3].as_ref()).is_none());
Run

pub fn get<Q>(&self, key: &Q) -> Option<&V>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Returns a reference to the value corresponding to the key.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
assert_eq!(*map.get([1, 2, 3].as_ref()).unwrap(), 'a');
Run

pub fn get_key_value<Q>(&self, key: &Q) -> Option<(&K, &V)>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Returns the key-value pair corresponding to the supplied key.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
assert_eq!(map.get_key_value([1, 2, 3].as_ref()).unwrap(), (&Box::from([1, 2, 3]), &'a'));
Run

pub fn get_mut<Q>(&mut self, key: &Q) -> Option<&mut V>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Returns a mutable reference to the value corresponding to the key.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
assert_eq!(map[[1, 2, 3].as_ref()], 'a');

*map.get_mut([1, 2, 3].as_ref()).unwrap() = 'b';
assert_eq!(map[[1, 2, 3].as_ref()], 'b');
Run

While an element from the tree is mutably referenced, no other operation on the tree can happen.

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();

let value = map.get_mut(&[1, 2, 3]).unwrap();
assert_eq!(*value, 'a');

assert_eq!(*map[[1, 2, 3].as_ref()], 'a');

*value = 'b';
drop(value);
Run

pub fn fuzzy<'a, 'b, Q>( &'a self, key: &'b Q, max_edit_dist: usize, ) -> Fuzzy<'a, 'b, K, V, PREFIX_LEN>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Makes a fuzzy search in the tree by key, returning all keys and values that are less than or equal to max_edit_dist.

This is done by using Levenshtein distance

Examples

use blart::TreeMap;

let mut map: TreeMap<_, _> = TreeMap::new();

map.insert(c"abc", 0);
map.insert(c"abd", 1);
map.insert(c"abdefg", 2);

let fuzzy: Vec<_> = map.fuzzy(c"ab", 2).collect();
assert_eq!(fuzzy, vec![(&c"abd", &1), (&c"abc", &0)]);
Run

pub fn fuzzy_mut<'a, 'b, Q>( &'a mut self, key: &'b Q, max_edit_dist: usize, ) -> FuzzyMut<'a, 'b, K, V, PREFIX_LEN>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Makes a fuzzy search in the tree by key, returning all keys and values that are less than or equal to max_edit_dist.

This is done by using Levenshtein distance

Examples

use blart::TreeMap;

let mut map: TreeMap<_, _> = TreeMap::new();

map.insert(c"abc", 0);
map.insert(c"abd", 1);
map.insert(c"abdefg", 2);

let fuzzy: Vec<_> = map.fuzzy_mut(c"ab", 2).collect();
assert_eq!(fuzzy, vec![(&c"abd", &mut 1), (&c"abc", &mut 0)]);
Run

pub fn fuzzy_keys<'a, 'b, Q>( &'a self, key: &'b Q, max_edit_dist: usize, ) -> FuzzyKeys<'a, 'b, K, V, PREFIX_LEN>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Makes a fuzzy search in the tree by key, returning all keys and values that are less than or equal to max_edit_dist

This is done by using Levenshtein distance

Examples

use blart::TreeMap;

let mut map: TreeMap<_, _> = TreeMap::new();

map.insert(c"abc", 0);
map.insert(c"abd", 1);
map.insert(c"abdefg", 2);

let fuzzy: Vec<_> = map.fuzzy_keys(c"ab", 2).collect();
assert_eq!(fuzzy, vec![&c"abd", &c"abc"]);
Run

pub fn fuzzy_values<'a, 'b, Q>( &'a self, key: &'b Q, max_edit_dist: usize, ) -> FuzzyValues<'a, 'b, K, V, PREFIX_LEN>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Makes a fuzzy search in the tree by key, returning all keys and values that are less than or equal to max_edit_dist.

This is done by using Levenshtein distance

Examples

use blart::TreeMap;

let mut map: TreeMap<_, _> = TreeMap::new();

map.insert(c"abc", 0);
map.insert(c"abd", 1);
map.insert(c"abdefg", 2);

let fuzzy: Vec<_> = map.fuzzy_values(c"ab", 2).collect();
assert_eq!(fuzzy, vec![&1, &0]);
Run

pub fn fuzzy_values_mut<'a, 'b, Q>( &'a mut self, key: &'b Q, max_edit_dist: usize, ) -> FuzzyValuesMut<'a, 'b, K, V, PREFIX_LEN>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Makes a fuzzy search in the tree by key, returning all keys and values that are less than or equal to max_edit_dist

This is done by using Levenshtein distance

Examples

use blart::TreeMap;

let mut map: TreeMap<_, _> = TreeMap::new();

map.insert(c"abc", 0);
map.insert(c"abd", 1);
map.insert(c"abdefg", 2);

let fuzzy: Vec<_> = map.fuzzy_values(c"ab", 2).collect();
assert_eq!(fuzzy, vec![&mut 1, &mut 0]);
Run

pub fn contains_key<Q>(&self, key: &Q) -> bool
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Returns true if the map contains a value for the specified key.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();

assert!(map.contains_key([1, 2, 3].as_ref()));
Run

pub fn first_key_value(&self) -> Option<(&K, &V)>

Returns the first key-value pair in the map. The key in this pair is the minimum key in the map.

If the tree is empty, returns None.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();

assert_eq!(map.first_key_value().unwrap(), (&[1, 2, 3].into(), &'a'));
Run

pub fn pop_first(&mut self) -> Option<(K, V)>

Removes and returns the first element in the map. The key of this element is the minimum key that was in the map.

If the tree is empty, returns None.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();

assert_eq!(map.pop_first().unwrap(), (Box::from([1, 2, 3]), 'a'));
Run

pub fn last_key_value(&self) -> Option<(&K, &V)>

Returns the last key-value pair in the map. The key in this pair is the maximum key in the map.

If the tree is empty, returns None.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
map.try_insert(Box::new([2, 3, 4]), 'b').unwrap();

assert_eq!(map.last_key_value().unwrap(), (&Box::from([2, 3, 4]), &'b'));
Run

pub fn pop_last(&mut self) -> Option<(K, V)>

Removes and returns the last element in the map. The key of this element is the maximum key that was in the map.

If the tree is empty, returns None.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
map.try_insert(Box::new([2, 3, 4]), 'b').unwrap();

assert_eq!(map.pop_last().unwrap(), (Box::from([2, 3, 4]), 'b'));
Run

fn init_tree(&mut self, key: K, value: V) -> NodePtr<PREFIX_LEN, LeafNode<K, V>>

fn apply_insert_point( &mut self, insert_point: InsertPoint<K, V, PREFIX_LEN>, key: K, value: V, ) -> InsertResult<'_, K, V, PREFIX_LEN>
where K: AsBytes,

fn apply_delete_point( &mut self, delete_point: DeletePoint<K, V, PREFIX_LEN>, ) -> DeleteResult<K, V, PREFIX_LEN>

pub fn insert(&mut self, key: K, value: V) -> Option<V>
where K: NoPrefixesBytes,

Insert a key-value pair into the map.

If the map did not have this key present, Ok(None) is returned.

If the map did have this key present, the value is updated, and the old value is returned.

Unlike try_insert, this function will not return an error, because the contract of the NoPrefixesBytes ensures that the given key type will never be a prefix of an existing value.

Examples

use blart::TreeMap;

let mut map = TreeMap::<u128, char>::new();

assert!(map.insert(123, 'a').is_none());
assert!(map.insert(234, 'b').is_none());
assert_eq!(map.insert(234, 'c'), Some('b'));

assert_eq!(map.len(), 2);
Run

pub fn try_insert( &mut self, key: K, value: V, ) -> Result<Option<V>, InsertPrefixError>
where K: AsBytes,

Inserts a key-value pair into the map.

If the map did not have this key present, Ok(None) is returned.

If the map did have this key present, the value is updated, and the old value is returned.

Errors

If the map has an existing key, such that the new key is a prefix of the existing key or vice versa, then it returns an error.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

assert!(map.try_insert(Box::new([1, 2, 3]), 'a').unwrap().is_none());
assert!(map.try_insert(Box::new([2, 3, 4]), 'b').unwrap().is_none());
// This function call errors because the key is a prefix of the existing key
assert!(map.try_insert(Box::new([2, 3, 4, 5]), 'c').is_err());
assert_eq!(map.try_insert(Box::new([2, 3, 4]), 'd').unwrap(), Some('b'));

assert_eq!(map.len(), 2);
Run

pub fn remove_entry<Q>(&mut self, key: &Q) -> Option<(K, V)>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Removes a key from the map, returning the stored key and value if the key was previously in the map.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
map.try_insert(Box::new([2, 3, 4]), 'b').unwrap();

assert_eq!(map.remove_entry([2, 3, 4].as_ref()).unwrap(), (Box::from([2, 3, 4]), 'b'))
Run

pub fn remove<Q>(&mut self, key: &Q) -> Option<V>
where K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Removes a key from the map, returning the value at the key if the key was previously in the map.

Examples

use blart::TreeMap;

let mut map = TreeMap::<Box<[u8]>, char>::new();

map.try_insert(Box::new([1, 2, 3]), 'a').unwrap();
map.try_insert(Box::new([2, 3, 4]), 'b').unwrap();

assert_eq!(map.remove([2, 3, 4].as_ref()).unwrap(), 'b');
assert_eq!(map.remove([2, 3, 4].as_ref()), None);
Run

pub fn into_keys(self) -> IntoKeys<K, V, PREFIX_LEN>

Creates a consuming iterator visiting all the keys, in sorted order. The map cannot be used after calling this. The iterator element type is K.

Examples

use blart::TreeMap;

let map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

let mut iter = map.into_keys();

assert_eq!(iter.next().unwrap(), 0);
assert_eq!(iter.next().unwrap(), 1);
assert_eq!(iter.next().unwrap(), 2);
assert_eq!(iter.next().unwrap(), 3);
assert_eq!(iter.next().unwrap(), 4);
assert_eq!(iter.next(), None);
Run

pub fn into_values(self) -> IntoValues<K, V, PREFIX_LEN>

Creates a consuming iterator visiting all the values, in order by key. The map cannot be used after calling this. The iterator element type is V.

Examples

use blart::TreeMap;

let map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

let mut iter = map.into_values();

assert_eq!(iter.next().unwrap(), 'd');
assert_eq!(iter.next().unwrap(), 'c');
assert_eq!(iter.next().unwrap(), 'b');
assert_eq!(iter.next().unwrap(), 'a');
assert_eq!(iter.next().unwrap(), 'z');
assert_eq!(iter.next(), None);
Run

pub fn iter(&self) -> TreeIterator<'_, K, V, PREFIX_LEN>

Gets an iterator over the entries of the map, sorted by key.

Examples

use blart::TreeMap;

let map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

let mut iter = map.iter();

assert_eq!(iter.next().unwrap(), (&0, &'d'));
assert_eq!(iter.next().unwrap(), (&1, &'c'));
assert_eq!(iter.next().unwrap(), (&2, &'b'));
assert_eq!(iter.next().unwrap(), (&3, &'a'));
assert_eq!(iter.next().unwrap(), (&4, &'z'));
assert_eq!(iter.next(), None);
Run

pub fn iter_mut(&mut self) -> TreeIteratorMut<'_, K, V, PREFIX_LEN>

Gets a mutable iterator over the entries of the map, sorted by key.

Examples

use blart::TreeMap;

let mut map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

for (_key, value) in map.iter_mut() {
    value.make_ascii_uppercase();
}

assert_eq!(map[&0], 'D');
assert_eq!(map[&1], 'C');
assert_eq!(map[&2], 'B');
assert_eq!(map[&3], 'A');
assert_eq!(map[&4], 'Z');
Run

pub fn keys(&self) -> Keys<'_, K, V, PREFIX_LEN>

Gets an iterator over the keys of the map, in sorted order.

Examples

use blart::TreeMap;

let map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

let mut iter = map.keys();

assert_eq!(iter.next().unwrap(), &0);
assert_eq!(iter.next().unwrap(), &1);
assert_eq!(iter.next().unwrap(), &2);
assert_eq!(iter.next().unwrap(), &3);
assert_eq!(iter.next().unwrap(), &4);
assert_eq!(iter.next(), None);
Run

pub fn values(&self) -> Values<'_, K, V, PREFIX_LEN>

Gets an iterator over the values of the map, in order by key.

Examples

use blart::TreeMap;

let map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

let mut iter = map.values();

assert_eq!(iter.next().unwrap(), &'d');
assert_eq!(iter.next().unwrap(), &'c');
assert_eq!(iter.next().unwrap(), &'b');
assert_eq!(iter.next().unwrap(), &'a');
assert_eq!(iter.next().unwrap(), &'z');
assert_eq!(iter.next(), None);
Run

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V, PREFIX_LEN>

Gets a mutable iterator over the values of the map, in order by key.

Examples

use blart::TreeMap;

let mut map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

for value in map.values_mut() {
    value.make_ascii_uppercase();
}

assert_eq!(map[&0], 'D');
assert_eq!(map[&1], 'C');
assert_eq!(map[&2], 'B');
assert_eq!(map[&3], 'A');
assert_eq!(map[&4], 'Z');
Run

pub fn prefix<'a, 'b>( &'a self, prefix: &'b [u8], ) -> Prefix<'a, 'b, K, V, PREFIX_LEN>
where K: AsBytes,

Gets an iterator over the entries of the map that start with prefix

Example

use blart::TreeMap;

let mut map = TreeMap::new();
map.insert(c"abcde", 0);
map.insert(c"abcdexxx", 0);
map.insert(c"abcdexxy", 0);
map.insert(c"abcdx", 0);
map.insert(c"abcx", 0);
map.insert(c"bx", 0);

let p: Vec<_> = map.prefix(c"abcde".to_bytes()).collect();

assert_eq!(p, vec![(&c"abcde", &0), (&c"abcdexxx", &0), (&c"abcdexxy", &0)]);
Run

pub fn prefix_mut<'a, 'b>( &'a mut self, prefix: &'b [u8], ) -> PrefixMut<'a, 'b, K, V, PREFIX_LEN>
where K: AsBytes,

Gets a mutable iterator over the entries of the map that start with prefix

Example

use blart::TreeMap;

let mut map = TreeMap::new();
map.insert(c"abcde", 0);
map.insert(c"abcdexxx", 0);
map.insert(c"abcdexxy", 0);
map.insert(c"abcdx", 0);
map.insert(c"abcx", 0);
map.insert(c"bx", 0);

let p: Vec<_> = map.prefix_mut(c"abcde".to_bytes()).collect();

assert_eq!(p, vec![(&c"abcde", &mut 0), (&c"abcdexxx", &mut 0), (&c"abcdexxy", &mut 0)]);
Run

pub fn prefix_keys<'a, 'b>( &'a self, prefix: &'b [u8], ) -> PrefixKeys<'a, 'b, K, V, PREFIX_LEN>
where K: AsBytes,

Gets an iterator over the keys of the map that start with prefix

Example

use blart::TreeMap;

let mut map = TreeMap::new();
map.insert(c"abcde", 0);
map.insert(c"abcdexxx", 0);
map.insert(c"abcdexxy", 0);
map.insert(c"abcdx", 0);
map.insert(c"abcx", 0);
map.insert(c"bx", 0);

let p: Vec<_> = map.prefix_keys(c"abcde".to_bytes()).collect();

assert_eq!(p, vec![&c"abcde", &c"abcdexxx", &c"abcdexxy"]);
Run

pub fn prefix_values<'a, 'b>( &'a self, prefix: &'b [u8], ) -> PrefixValues<'a, 'b, K, V, PREFIX_LEN>
where K: AsBytes,

Gets an iterator over the values of the map that start with prefix

Example

use blart::TreeMap;

let mut map = TreeMap::new();
map.insert(c"abcde", 0);
map.insert(c"abcdexxx", 1);
map.insert(c"abcdexxy", 2);
map.insert(c"abcdx", 3);
map.insert(c"abcx", 4);
map.insert(c"bx", 5);

let p: Vec<_> = map.prefix_values(c"abcde".to_bytes()).collect();

assert_eq!(p, vec![&0, &1, &2]);
Run

pub fn prefix_values_mut<'a, 'b>( &'a mut self, prefix: &'b [u8], ) -> PrefixValuesMut<'a, 'b, K, V, PREFIX_LEN>
where K: AsBytes,

Gets a mutable iterator over the values of the map that start with prefix

Example

use blart::TreeMap;

let mut map = TreeMap::new();
map.insert(c"abcde", 0);
map.insert(c"abcdexxx", 1);
map.insert(c"abcdexxy", 2);
map.insert(c"abcdx", 3);
map.insert(c"abcx", 4);
map.insert(c"bx", 5);

let p: Vec<_> = map.prefix_values(c"abcde".to_bytes()).collect();

assert_eq!(p, vec![&mut 0, &mut 1, &mut 2]);
Run

pub fn len(&self) -> usize

Returns the number of elements in the map.

Examples

use blart::TreeMap;

let map: TreeMap<_, char> = ['d', 'c', 'b', 'a', 'z'].into_iter()
    .enumerate()
    .collect();

assert_eq!(map.len(), 5);
Run

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Examples

use blart::TreeMap;

let map = TreeMap::<Box<[u8]>, ()>::new();
assert!(map.is_empty());
Run

impl<K, V, const PREFIX_LEN: usize> TreeMap<K, V, PREFIX_LEN>

pub fn try_entry( &mut self, key: K, ) -> Result<Entry<'_, K, V, PREFIX_LEN>, InsertPrefixError>
where K: AsBytes,

Tries to get the given key’s corresponding entry in the map for in-place manipulation.

pub fn try_entry_ref<'a, 'b, Q>( &'a mut self, key: &'b Q, ) -> Result<EntryRef<'a, 'b, K, V, Q, PREFIX_LEN>, InsertPrefixError>
where K: AsBytes + Borrow<Q> + From<&'b Q>, Q: AsBytes + ?Sized,

Tries to get the given key’s corresponding entry in the map for in-place manipulation.

pub fn entry(&mut self, key: K) -> Entry<'_, K, V, PREFIX_LEN>
where K: NoPrefixesBytes,

Gets the given key’s corresponding entry in the map for in-place manipulation.

pub fn entry_ref<'a, 'b, Q>( &'a mut self, key: &'b Q, ) -> EntryRef<'a, 'b, K, V, Q, PREFIX_LEN>
where K: NoPrefixesBytes + Borrow<Q> + From<&'b Q>, Q: NoPrefixesBytes + ?Sized,

Gets the given key’s corresponding entry in the map for in-place manipulation.