Function blart::insert_unchecked

pub unsafe fn insert_unchecked<V>(
    root: OpaqueNodePtr<V>, 
    key: Box<[u8]>, 
    value: V
) -> Result<OpaqueNodePtr<V>, InsertError>

Insert the given key-value pair into the tree.

Returns either a pointer to the new tree root or an error.

Errors

Returns a InsertError::EmptyKey if the given key is an empty array.

Returns a InsertError::PrefixKey if the given key is a prefix of another key that exists in the trie. Or if the given key is prefixed by an existing key in the trie.

Safety

The root OpaqueNodePtr must be a unique pointer to the underlyin

This function cannot be called concurrently to any reads or writes of the root node or any child node of root. This function will arbitrarily read or write to any child in the given tree.