Module blart::nightly_rust_apis

Module containing copies of Rust standard library unstable functions for use outside of the nightly distribution.

Functions

hasher_write_length_prefix

Writes a length prefix into this hasher, as part of being prefix-free.

maybe_uninit_slice_as_ptr

Gets a pointer to the first element of the array.

maybe_uninit_slice_assume_init_mut⚠

Assuming all the elements are initialized, get a mutable slice to them.

maybe_uninit_slice_assume_init_ref⚠

Assuming all the elements are initialized, get a slice to them.

maybe_uninit_uninit_array

Create a new array of MaybeUninit<T> items, in an uninitialized state.

non_null_get_unchecked_mut⚠

Returns a raw pointer to an element, without doing bounds checking.

non_null_slice_from_raw_parts

Creates a non-null raw slice from a thin pointer and a length.