Module blart::nodes::operations::insert

Structs

InsertPoint

This struct contains the results from searching for an insert point for a new node in the tree.

InsertPrefixError

Attempted to insert a key which was a prefix of an existing key in the tree.

InsertResult

The results of a successful tree insert

Enums

InsertSearchResultType

The type of insert

Functions

search_for_insert_point⚠

Perform an iterative search for the insert point for the given key, starting at the given root node.