Enum blart::nodes::operations::insert::InsertSearchResultType

pub enum InsertSearchResultType<K, V, const PREFIX_LEN: usize> {
    MismatchPrefix {
        mismatch: Mismatch<K, V, PREFIX_LEN>,
        mismatched_inner_node_ptr: OpaqueNodePtr<K, V, PREFIX_LEN>,
    },
    SplitLeaf {
        leaf_node_ptr: NodePtr<PREFIX_LEN, LeafNode<K, V>>,
        new_key_bytes_used: usize,
    },
    Exact {
        leaf_node_ptr: NodePtr<PREFIX_LEN, LeafNode<K, V>>,
    },
    IntoExisting {
        inner_node_ptr: OpaqueNodePtr<K, V, PREFIX_LEN>,
    },
}

The type of insert

Variants

MismatchPrefix

An insert where an inner node had a differing prefix from the key.

This insert type will create a new inner node with the portion of the prefix that did match, and update the existing inner node

Fields

mismatch: Mismatch<K, V, PREFIX_LEN>

Data about the matching if the prefix

mismatched_inner_node_ptr: OpaqueNodePtr<K, V, PREFIX_LEN>

A pointer to the inner node which had a mismatched prefix

SplitLeaf

An insert where the node to be added matched all the way up to a leaf node.

This insert type will create a new inner node, and assign the existing leaf and the new leaf as children to that node.

Fields

leaf_node_ptr: NodePtr<PREFIX_LEN, LeafNode<K, V>>

A pointer to the leaf node that will be split

new_key_bytes_used: usize

Exact

Exact match of the leaf was found

This insert type will replace the older leaf with a new one

Fields

leaf_node_ptr: NodePtr<PREFIX_LEN, LeafNode<K, V>>

A pointer to the leaf node that will be split

IntoExisting

An insert where the search terminated at an existing inner node that did not have a child with the key byte.

If the inner node is full, it will be grown to the next largest size.

Fields

inner_node_ptr: OpaqueNodePtr<K, V, PREFIX_LEN>

A pointer to the existing inner node which will be updated to contain the new child leaf node