Module blart::nodes::representation

Trie node representation

Re-exports

pub use self::iterators::*;

Modules

iterators 🔒

private 🔒

Structs

Header

The common header for all inner nodes

InnerNode48

Node that references between 17 and 49 children

InnerNode256

Node that references between 49 and 256 children

InnerNodeCompressed

Node type that has a compact representation for key bytes and children pointers.

LeafNode

Node that contains a single leaf value.

NodePtr

A pointer to a Node.

OpaqueNodePtr

An opaque pointer to a Node.

OpaqueValue 🔒

A placeholder type that has the required amount of alignment.

RestrictedNodeIndex

A restricted index only valid from 0 to LIMIT - 1.

TryFromByteError

The error type returned when attempting to construct an index outside the accepted range of a PartialNodeIndex.

Enums

ConcreteNodePtr

An enum that encapsulates pointers to every type of Node

NodeType

The representation of inner nodes

Constants

NUM_PREFIX_BYTES

The number of bytes stored for path compression in the node header.

Traits

InnerNode

Common methods implemented by all inner node.

Node

All nodes which contain a runtime tag that validates their type.

Type Definitions

InnerNode4

Node that references between 2 and 4 children

InnerNode16

Node that references between 5 and 16 children