Type Alias blart::nodes::representation::inner_node_compressed::InnerNode16

pub type InnerNode16<K, V, const PREFIX_LEN: usize> = InnerNodeCompressed<K, V, PREFIX_LEN, 16>;

Node that references between 5 and 16 children

Aliased Type

struct InnerNode16<K, V, const PREFIX_LEN: usize> {
    pub header: Header<PREFIX_LEN>,
    pub keys: [MaybeUninit<u8>; 16],
    pub child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; 16],
}

Fields
header: Header<PREFIX_LEN>

The common node fields.

keys: [MaybeUninit<u8>; 16]

An array that contains single key bytes in the same index as the child_pointers array contains the matching child tree.

This array will only be initialized for the first header.num_children values.

child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; 16]

An array that contains the child data.

This array will only be initialized for the first header.num_children values.