Struct blart::nodes::representation::InnerNode48

pub struct InnerNode48<K, V> {
    pub header: Header,
    pub child_indices: [RestrictedNodeIndex<48>; 256],
    pub child_pointers: [MaybeUninit<OpaqueNodePtr<K, V>>; 48],
}

Node that references between 17 and 49 children

Fields
header: Header

The common node fields.

child_indices: [RestrictedNodeIndex<48>; 256]

An array that maps key bytes (as the index) to the index value in the child_pointers array.

All the child_indices values are guaranteed to be PartialNodeIndex<48>::EMPTY when the node is constructed.

child_pointers: [MaybeUninit<OpaqueNodePtr<K, V>>; 48]

For each element in this array, it is assumed to be initialized if there is a index in the child_indices array that points to it

Implementations

impl<K, V> InnerNode48<K, V>

pub fn empty() -> Self

Create an empty Node48.

pub fn initialized_child_pointers(&self) -> &[OpaqueNodePtr<K, V>]

Return the initialized portions of the child pointer array.