blart::nodes::representation
Struct LeafNode

#[repr(align(8))]
pub struct LeafNode<K, V, const PREFIX_LEN: usize> {
    value: V,
    key: K,
    pub(crate) previous: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
    pub(crate) next: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
}

Node that contains a single leaf value.

Fields
value: V

The leaf value.

key: K

The full key that the value was stored with.

previous: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>

Pointer to the previous leaf node in the trie. If the value is None, then this is the first leaf.

next: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>

Pointer to the next leaf node in the trie. If the value is None, then this is the last leaf.

Implementations

impl<K, V, const PREFIX_LEN: usize> LeafNode<K, V, PREFIX_LEN>
where K: AsBytes,

pub unsafe fn insert_after( this_ptr: NodePtr<PREFIX_LEN, Self>, previous_sibling_ptr: NodePtr<PREFIX_LEN, Self>, )

Insert the leaf node pointed to by this_ptr into the linked list that previous_sibling_ptr belongs to, placing the “this” leaf node after the “previous sibling” in the list.

Safety

This function requires that no other operation is concurrently modifying or reading the this_ptr leaf node, the previous_sibling_ptr leaf node, and the sibling leaf node of previous_sibling_ptr.

pub unsafe fn insert_before( this_ptr: NodePtr<PREFIX_LEN, Self>, next_sibling_ptr: NodePtr<PREFIX_LEN, Self>, )

Insert the leaf node pointed to by this_ptr into the linked list that next_sibling_ptr belongs to, placing the “this” leaf node before the “next sibling” in the list.

Safety

This function requires that no other operation is concurrently modifying or reading the this_ptr leaf node, the next_sibling_ptr leaf node, and the sibling leaf node of next_sibling_ptr.

pub unsafe fn replace(this_ptr: NodePtr<PREFIX_LEN, Self>, old_leaf: &mut Self)

Insert the leaf node pointed to by this_ptr into the linked list position that old_leaf currently occupies, and then remove old_leaf from the linked list.

Safety

This function requires that no other operation is concurrently modifying or reading the this_ptr leaf node and the sibling leaf nodes of the old_leaf.

impl<const PREFIX_LEN: usize, K, V> LeafNode<K, V, PREFIX_LEN>

pub fn with_no_siblings(key: K, value: V) -> Self

Create a new leaf node with the given value and no siblings.

pub fn key_ref(&self) -> &K

Returns a shared reference to the key contained by this leaf node

pub fn value_ref(&self) -> &V

Returns a shared reference to the value contained by this leaf node

pub fn value_mut(&mut self) -> &mut V

Returns a mutable reference to the value contained by this leaf node

pub fn entry_ref(&self) -> (&K, &V)

Return shared references to the key and value contained by this leaf node

pub fn entry_mut(&mut self) -> (&mut K, &mut V)

Return mutable references to the key and value contained by this leaf node

pub fn into_entry(self) -> (K, V)

Consume the leaf node and return a tuple of the key and value

pub fn matches_full_key(&self, possible_key: &[u8]) -> bool
where K: AsBytes,

Check that the provided full key is the same one as the stored key.

pub unsafe fn remove_self(this_ptr: NodePtr<PREFIX_LEN, Self>)

This function removes this leaf node from its linked list.

Safety

This function requires that no other operation is concurrently modifying or reading the this_ptr leaf node and its sibling leaf nodes.

pub fn clone_without_siblings(&self) -> Self
where K: Clone, V: Clone,

Create a copy of this leaf node with the sibling references removed.