Trait blart::nodes::representation::InnerNode

pub trait InnerNode<const PREFIX_LEN: usize>: Node<PREFIX_LEN> + Sized {
    type GrownNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>;
    type ShrunkNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>;
    type Iter<'a>: Iterator<Item = (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>)> + DoubleEndedIterator + FusedIterator
       where Self: 'a;

    // Required methods
    fn from_header(header: Header<PREFIX_LEN>) -> Self;
    fn header(&self) -> &Header<PREFIX_LEN>;
    fn lookup_child(
        &self,
        key_fragment: u8,
    ) -> Option<OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>>;
    fn write_child(
        &mut self,
        key_fragment: u8,
        child_pointer: OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>,
    );
    fn remove_child(
        &mut self,
        key_fragment: u8,
    ) -> Option<OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>>;
    fn grow(&self) -> Self::GrownNode;
    fn shrink(&self) -> Self::ShrunkNode;
    fn iter(&self) -> Self::Iter<'_>;
    fn range(
        &self,
        bound: impl RangeBounds<u8>,
    ) -> impl DoubleEndedIterator<Item = (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>)> + FusedIterator;
    fn min(&self) -> (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>);
    fn max(&self) -> (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>);
    fn deep_clone(&self) -> NodePtr<PREFIX_LEN, Self>
       where Self::Key: Clone,
             Self::Value: Clone;

    // Provided methods
    fn empty() -> Self { ... }
    fn from_prefix(prefix: &[u8], prefix_len: usize) -> Self { ... }
    fn is_full(&self) -> bool { ... }
    fn match_prefix(
        &self,
        key: &[u8],
        current_depth: usize,
    ) -> MatchPrefixResult<Self::Key, Self::Value, PREFIX_LEN>
       where Self::Key: AsBytes { ... }
    fn read_full_prefix(
        &self,
        current_depth: usize,
    ) -> (&[u8], Option<NodePtr<PREFIX_LEN, LeafNode<Self::Key, Self::Value>>>)
       where Self::Key: AsBytes { ... }

}

Common methods implemented by all inner node.

Required Associated Types

type GrownNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>

The type of the next larger node type.

type ShrunkNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>

The type of the next smaller node type.

type Iter<'a>: Iterator<Item = (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>)> + DoubleEndedIterator + FusedIterator where Self: 'a

The type of the iterator over all children of the inner node

Required Methods

fn from_header(header: Header<PREFIX_LEN>) -> Self

Create a new InnerNode using a Header

fn header(&self) -> &Header<PREFIX_LEN>

Get the Header from the InnerNode

fn lookup_child( &self, key_fragment: u8, ) -> Option<OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>>

Search through this node for a child node that corresponds to the given key fragment.

fn write_child( &mut self, key_fragment: u8, child_pointer: OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>, )

Write a child pointer with key fragment to this inner node.

If the key fragment already exists in the node, overwrite the existing child pointer.

Panics

Panics when the node is full.

fn remove_child( &mut self, key_fragment: u8, ) -> Option<OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>>

Attempt to remove a child pointer at the key fragment from this inner node.

If the key fragment does not exist in this node, return None.

fn grow(&self) -> Self::GrownNode

Grow this node into the next larger class, copying over children and prefix information.

fn shrink(&self) -> Self::ShrunkNode

Shrink this node into the next smaller class, copying over children and prefix information.

Panics

Panics if the new, smaller node size does not have enough capacity to hold all the children.

fn iter(&self) -> Self::Iter<'_>

Create an iterator over all (key bytes, child pointers) in this inner node.

fn range( &self, bound: impl RangeBounds<u8>, ) -> impl DoubleEndedIterator<Item = (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>)> + FusedIterator

Create an iterator over a subset of (key bytes, child pointers), using the given bound as a restriction on the set of key bytes.

fn min(&self) -> (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>)

Returns the minimum child pointer from this node and it’s key

Safety

Since this is a InnerNode we assume that the we have at least one child, (more strictly we have 2, because with one child the node would have collapsed) so in this way we can avoid the Option. This is safe because if we had no children this current node should have been deleted.

fn max(&self) -> (u8, OpaqueNodePtr<Self::Key, Self::Value, PREFIX_LEN>)

Returns the maximum child pointer from this node and it’s key

Safety

Since this is a InnerNode we assume that the we have at least one child, (more strictly we have 2, because with one child the node would have collapsed) so in this way we can avoid the Option. This is safe because if we had, no children this current node should have been deleted.

fn deep_clone(&self) -> NodePtr<PREFIX_LEN, Self>
where Self::Key: Clone, Self::Value: Clone,

Deep clones the inner node by allocating memory to a new one

Provided Methods

fn empty() -> Self

Create an empty InnerNode, with no children and no prefix

fn from_prefix(prefix: &[u8], prefix_len: usize) -> Self

Create a new InnerNode using prefix as the node prefix and prefix_len as the node prefix length and

This is done because when a prefix mismatch happens the length of the mismatch can be grater or equal to prefix size, since we search for the first child of the node to recreate the prefix, that’s why we don’t use prefix.len() as the node prefix length

fn is_full(&self) -> bool

Returns true if this node has no more space to store children.

fn match_prefix( &self, key: &[u8], current_depth: usize, ) -> MatchPrefixResult<Self::Key, Self::Value, PREFIX_LEN>
where Self::Key: AsBytes,

Compares the compressed path of a node with the key and returns the number of equal bytes.

Safety

current_depth > key len

fn read_full_prefix( &self, current_depth: usize, ) -> (&[u8], Option<NodePtr<PREFIX_LEN, LeafNode<Self::Key, Self::Value>>>)
where Self::Key: AsBytes,

Read the prefix as a whole, by reconstructing it if necessary from a leaf

Object Safety

This trait is not object safe.