Module blart::nodes::visitor::tree_stats

Structs

InnerNodeStats

Statistics for inner nodes

LeafStats

Statistics for inner nodes

TreeStats

Collection of stats about the number of nodes types present in a tree

TreeStatsCollector

A visitor of the radix tree which collects statistics about the tree, like how many inner nodes of each type, how many leaves