Struct blart::nodes::visitor::tree_stats::TreeStatsCollector

pub struct TreeStatsCollector;

A visitor of the radix tree which collects statistics about the tree, like how many inner nodes of each type, how many leaves

Implementations

impl TreeStatsCollector

pub unsafe fn collect<K, V>(root: OpaqueNodePtr<K, V>) -> TreeStatswhere K: AsBytes,

Run the tree stats collection on the given root node, then return the accumalated stats.

Safety

For the duration of this function, the given node and all its children nodes must not get mutated.

pub unsafe fn count_leaf_nodes<K, V>(root: OpaqueNodePtr<K, V>) -> usize

Iterate through the given tree and return the number of leaf nodes.

Safety

For the duration of this function, the given node and all its children nodes must not get mutated.