blart
Module raw

Trie node representation and manipulation

Modules

iterator 🔒

operations 🔒

Trie node lookup and manipulation

representation 🔒

Trie node representation

visitor

Utilities for inspecting the trie structure.

Structs

AttemptOptimisticPrefixMatch

This struct represents a successful match against a prefix using the InnerNode::attempt_pessimistic_match_prefix function.

ExplicitMismatch

Represents a prefix mismatch when looking at the entire prefix, including in cases where it is read from a child leaf node.

InnerNode48

Node that references between 17 and 49 children

InnerNode256

Node that references between 49 and 256 children

InnerNodeCompressed

Node type that has a compact representation for key bytes and children pointers.

LeafNode

Node that contains a single leaf value.

NodePtr

A pointer to a Node.

OpaqueNodePtr

An opaque pointer to a Node.

OptimisticMismatch

Represents a prefix mismatch when looking only at the prefix content present in an InnerNode header.

PessimisticMismatch

Represents a prefix mismatch when looking only at the prefix content present in an InnerNode header.

PrefixMatch

This struct represents a successful match against a prefix using either the InnerNode::optimistic_match_prefix or InnerNode::match_full_prefix functions.

RestrictedNodeIndex

A restricted index only valid from 0 to LIMIT - 1.

TryFromByteError

The error type returned when attempting to construct an index outside the accepted range of a PartialNodeIndex.

Enums

ConcreteInnerNodePtr

An enum that encapsulates pointers to every type of InnerNode

ConcreteNodePtr

An enum that encapsulates pointers to every type of Node

NodeType

The representation of inner nodes

Traits

InnerNode

Common methods implemented by all inner node.

Node

All nodes which contain a runtime tag that validates their type.

Type Aliases

InnerNode4

Node that references between 2 and 4 children

InnerNode16

Node that references between 5 and 16 children

InnerNodeCompressedIter

Iterator type for an InnerNodeCompressed