blart::raw::operations::deallocate

Function deallocate_inner_node

unsafe fn deallocate_inner_node<K, V, N, A: Allocator, const PREFIX_LEN: usize>(
    stack: &mut Vec<OpaqueNodePtr<K, V, PREFIX_LEN>>,
    inner_ptr: NodePtr<PREFIX_LEN, N>,
    filter_children: for<'a> fn(&'a OpaqueNodePtr<K, V, PREFIX_LEN>) -> bool,
    alloc: &A,
)
where
    N: InnerNode<PREFIX_LEN, Key = K, Value = V>,

This function will read an InnerNode and place all children that satisfy the filter_children predicate on the stack. Then it will deallocate the given InnerNode.

Safety

This function must only be called once for this inner node

This function cannot be called concurrently with any read of modification of the trie.