blart::raw::operations::insert

Function search_for_insert_point

pub unsafe fn search_for_insert_point<K, V, const PREFIX_LEN: usize>(
    root: OpaqueNodePtr<K, V, PREFIX_LEN>,
    key_bytes: &[u8],
) -> Result<InsertPoint<K, V, PREFIX_LEN>, InsertPrefixError>
where
    K: AsBytes,

Perform an iterative search for the insert point for the given key, starting at the given root node.

Safety

This function cannot be called concurrently to any writes of the root node or any child node of root. This function will arbitrarily read to any child in the given tree.

Errors

If the given key is a prefix of an existing key, this function will return an error.