blart::raw::operations::lookup

Function search_unchecked

pub unsafe fn search_unchecked<K, V, const PREFIX_LEN: usize>(
    root: OpaqueNodePtr<K, V, PREFIX_LEN>,
    key_bytes: &[u8],
) -> Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>
where
    K: AsBytes,

Search in the given tree for the value stored with the given key.

Safety

This function cannot be called concurrently with any mutating operation on root or any child node of root. This function will arbitrarily read to any child in the given tree.