blart::raw

Module representation

Trie node representation

Re-exports

pub use header::*;

pub use inner_node_direct::*;

pub use inner_node_indirect::*;

pub use inner_node_sorted::*;

pub use leaf_node::*;

pub use pointers::*;

Modules

header 🔒

Inner node header type

inner_node_direct 🔒

inner_node_indirect 🔒

inner_node_sorted 🔒

leaf_node 🔒

pointers 🔒

This module contains definitions of pointers types used in the tree.

private 🔒

Structs

AttemptOptimisticPrefixMatch

This struct represents a successful match against a prefix using the InnerNodeCommon::attempt_pessimistic_match_prefix function.

ExplicitMismatch

Represents a prefix mismatch when looking at the entire prefix, including in cases where it is read from a child leaf node.

OptimisticMismatch

Represents a prefix mismatch when looking only at the prefix content present in an InnerNode header.

PessimisticMismatch

Represents a prefix mismatch when looking only at the prefix content present in an InnerNode header.

PrefixMatch

This struct represents a successful match against a prefix using either the InnerNodeCommon::optimistic_match_prefix or InnerNodeCommon::match_full_prefix functions.

TreePathSearch

This type is used to track the parent and grandparent when searching down the tree.

Enums

NodeType

The representation of inner nodes

TreePath

This enum represents different kinds of tree paths pointing to a leaf node.

Traits

InnerNode

Common methods implemented by all inner nodes that will be used in the tree.

InnerNodeCommon

This trait contains all the functions/types that every inner node must implement.

Node

All nodes which contain a runtime tag that validates their type.

Functions

assert_valid_range_bounds 🔒

Check that the given range bounds are valid, if not then panic.

Type Aliases

NodePrefix

This type represents the contents of an InnerNode prefix, either read directly from the prefix or fetched from a leaf node descendant.