blart::raw::representation::inner_node_48

Struct InnerNode48

#[repr(C, align(8))]
pub struct InnerNode48<K, V, const PREFIX_LEN: usize> {
    pub header: Header<PREFIX_LEN>,
    pub child_indices: [RestrictedNodeIndex<48>; 256],
    pub child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; 48],
}

Node that references between 17 and 49 children

Fields
header: Header<PREFIX_LEN>

The common node fields.

child_indices: [RestrictedNodeIndex<48>; 256]

An array that maps key bytes (as the index) to the index value in the child_pointers array.

All the child_indices values are guaranteed to be PartialNodeIndex<48>::EMPTY when the node is constructed.

child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; 48]

For each element in this array, it is assumed to be initialized if there is a index in the child_indices array that points to it

Implementations

impl<K, V, const PREFIX_LEN: usize> InnerNode48<K, V, PREFIX_LEN>

pub fn initialized_child_pointers(&self) -> &[OpaqueNodePtr<K, V, PREFIX_LEN>]

Return the initialized portions of the child pointer array.