blart::raw::representation::inner_node_indirect

Struct InnerNodeIndirect

#[repr(C, align(8))]
pub struct InnerNodeIndirect<K, V, const PREFIX_LEN: usize, const SIZE: usize> {
    pub header: Header<PREFIX_LEN>,
    pub child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; SIZE],
    pub child_indices: [Option<NonMaxIndex>; 256],
}

Inner node type that uses a direct lookup for the key byte and uses the resulting value to lookup the child pointer.

It is “indirect” in the sense that we use the value of the key byte to get an index that lets us finally look up the child pointer.

Fields
header: Header<PREFIX_LEN>

The common node fields.

child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; SIZE]

For each element in this array, it is assumed to be initialized if there is a index in the child_indices array that points to it

child_indices: [Option<NonMaxIndex>; 256]

An array that maps key bytes (as the index) to the index value in the child_pointers array.

All the child_indices values are guaranteed to be None when the node is constructed.

Implementations

impl<K, V, const PREFIX_LEN: usize, const SIZE: usize> InnerNodeIndirect<K, V, PREFIX_LEN, SIZE>

pub fn initialized_child_pointers(&self) -> &[OpaqueNodePtr<K, V, PREFIX_LEN>]

Return the initialized portions of the child pointer array.