blart::raw::representation::inner_node_indirect

Type Alias InnerNode48

pub type InnerNode48<K, V, const PREFIX_LEN: usize> = InnerNodeIndirect<K, V, PREFIX_LEN, 48>;

Node that references between 17 and 49 children.

Aliased Type

#[repr(C, align(8))]
pub struct InnerNode48<K, V, const PREFIX_LEN: usize> {
    pub header: Header<PREFIX_LEN>,
    pub child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; 48],
    pub child_indices: [Option<NonMaxIndex>; 256],
}

Fields
header: Header<PREFIX_LEN>

The common node fields.

child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; 48]

For each element in this array, it is assumed to be initialized if there is a index in the child_indices array that points to it

child_indices: [Option<NonMaxIndex>; 256]

An array that maps key bytes (as the index) to the index value in the child_pointers array.

All the child_indices values are guaranteed to be None when the node is constructed.