blart::raw

Struct InnerNodeSorted

#[repr(C, align(8))]
pub struct InnerNodeSorted<K, V, const PREFIX_LEN: usize, const SIZE: usize> {
    pub header: Header<PREFIX_LEN>,
    pub child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; SIZE],
    pub keys: [u8; SIZE],
}

Node type that has a compact representation for key bytes and children pointers.

The “sorted” aspect describes how the key bytes are stored. Key byte lookup is done via a linear search over an array, then using the index of a match to grab the corresponding child pointer from a separate array.

Fields
header: Header<PREFIX_LEN>

The common node fields.

child_pointers: [MaybeUninit<OpaqueNodePtr<K, V, PREFIX_LEN>>; SIZE]

An array that contains the child data.

This array will only be initialized for the first header.num_children values.

keys: [u8; SIZE]

An array that contains single key bytes in the same index as the child_pointers array contains the matching child tree.

This array will only be initialized for the first header.num_children values.

Implementations

impl<K, V, const PREFIX_LEN: usize, const SIZE: usize> InnerNodeSorted<K, V, PREFIX_LEN, SIZE>

pub fn initialized_portion(&self) -> (&[u8], &[OpaqueNodePtr<K, V, PREFIX_LEN>])

Return the initialized portions of the keys and child pointer arrays.

pub unsafe fn write_child_unchecked( &mut self, key_fragment: u8, child_pointer: OpaqueNodePtr<K, V, PREFIX_LEN>, )

Writes a child to the node without bounds check or order

Safety

This functions assumes that the write is gonna be inbound (i.e the check for a full node is done previously to the call of this function)

unsafe fn write_child_at( &mut self, idx: usize, key_fragment: u8, child_pointer: OpaqueNodePtr<K, V, PREFIX_LEN>, )

Writes a child to the node without bounds check or order

Safety

idx must be less than SIZE so that array accesses are in-bounds

fn change_block_size<const NEW_SIZE: usize>( &self, ) -> InnerNodeSorted<K, V, PREFIX_LEN, NEW_SIZE>

Grows or shrinks the node

pub(super) fn inner_iter(&self) -> InnerNodeSortedIter<'_, K, V, PREFIX_LEN>

Get an iterator over the keys and values of the node

fn lookup_child_index(&self, key_fragment: u8) -> Option<usize>

fn default_lookup_child_index(&self, key_fragment: u8) -> Option<usize>

fn n16_lookup_child_index(&self, key_fragment: u8) -> Option<usize>

fn find_write_point(&self, key_fragment: u8) -> WritePoint

fn n16_find_write_point(&self, key_fragment: u8) -> WritePoint

fn default_find_write_point(&self, key_fragment: u8) -> WritePoint