blart::raw

Trait InnerNode

pub trait InnerNode<const PREFIX_LEN: usize>:
    Node<PREFIX_LEN>
    + Sized
    + Debug
    + InnerNodeCommon<Self::Key, Self::Value, PREFIX_LEN> {
    type GrownNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>;
    type ShrunkNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>;

    // Required methods
    fn grow(&self) -> Self::GrownNode;
    fn shrink(&self) -> Self::ShrunkNode;

    // Provided method
    fn is_full(&self) -> bool { ... }
}

Common methods implemented by all inner nodes that will be used in the tree.

To contrast with InnerNodeCommon, this trait:

Incorporates a set value of NodeType, accessible via Node::TYPE.

Specifies where the inner node sits in the “growth order”, via InnerNode::GrownNode and InnerNode::ShrunkNode.

Required Associated Types

type GrownNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>

The type of the next larger node type.

type ShrunkNode: InnerNode<PREFIX_LEN, Key = Self::Key, Value = Self::Value>

The type of the next smaller node type.

Required Methods

fn grow(&self) -> Self::GrownNode

Grow this node into the next larger class, copying over children and prefix information.

fn shrink(&self) -> Self::ShrunkNode

Shrink this node into the next smaller class, copying over children and prefix information.

Panics

Panics if the new, smaller node size does not have enough capacity to hold all the children.

Provided Methods

fn is_full(&self) -> bool

Returns true if this node has no more space to store children.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.