blart::raw::visitor

Function bytes_display_fmt

pub fn bytes_display_fmt(value: &impl AsBytes, f: &mut Formatter<'_>) -> Result

Print the byte representation of the given value as returned by AsBytes.

This function can be used to override the formatting of key or value types when printing with DotPrinter.

Errors

This function returns an error if the provided fmt::Formatter returns Err.