blart::raw::visitor

Module pretty_printer

Structs

DotPrinter

A visitor of the radix trie that will print the tree in “dot” notation.

DotPrinterSettings

Settings which customize the output of the DotPrinter visitor.

OverrideDisplay 🔒

Functions

bytes_display_fmt

Print the byte representation of the given value as returned by AsBytes.

debug_as_display_fmt

Print a given value using the Debug::fmt implementation.

null_display_fmt

Print the string “[null]”.