blart::raw::visitor::pretty_printer

Struct DotPrinter

pub struct DotPrinter<O: Write, K, V> {
    output: O,
    next_id: usize,
    settings: DotPrinterSettings,
    key_fmt: fn(&K, &mut Formatter<'_>) -> Result,
    value_fmt: fn(&V, &mut Formatter<'_>) -> Result,
}

A visitor of the radix trie that will print the tree in “dot” notation.

See ‘DOT Language | Graphviz’ for information about syntax and example of the language.

Fields
output: Onext_id: usizesettings: DotPrinterSettingskey_fmt: fn(&K, &mut Formatter<'_>) -> Resultvalue_fmt: fn(&V, &mut Formatter<'_>) -> Result
Implementations

impl<O: Write, K, V> DotPrinter<O, K, V>

pub fn print<const PREFIX_LEN: usize, A: Allocator>( output: O, tree: &TreeMap<K, V, PREFIX_LEN, A>, settings: DotPrinterSettings, ) -> Option<Result<()>>
where K: Display, V: Display,

Write the dot-format of the given tree to the given output.

pub fn print_with_fmt<const PREFIX_LEN: usize, A: Allocator>( output: O, tree: &TreeMap<K, V, PREFIX_LEN, A>, settings: DotPrinterSettings, key_fmt: fn(&K, &mut Formatter<'_>) -> Result, value_fmt: fn(&V, &mut Formatter<'_>) -> Result, ) -> Option<Result<()>>

Write the dot-format of the given tree to the given output.

pub(crate) unsafe fn print_tree<const PREFIX_LEN: usize>( output: O, tree: &OpaqueNodePtr<K, V, PREFIX_LEN>, settings: DotPrinterSettings, key_fmt: fn(&K, &mut Formatter<'_>) -> Result, value_fmt: fn(&V, &mut Formatter<'_>) -> Result, ) -> Result<()>

Write the dot-format of the given tree to the given output.

Safety

For the duration of this function, the given node and all its children nodes must not get mutated.

fn output_prelude(&mut self) -> Result<()>

fn output_epilogue(&mut self) -> Result<()>

fn get_id(&mut self) -> usize

fn write_inner_node<N, const PREFIX_LEN: usize>( &mut self, inner_node: &N, ) -> Result<usize>
where N: InnerNode<PREFIX_LEN, Key = K, Value = V>,