blart::raw::visitor

Struct WellFormedChecker

pub struct WellFormedChecker<K, V, const PREFIX_LEN: usize> {
    current_key_prefix: Vec<u8>,
    seen_nodes: HashMap<OpaqueNodePtr<K, V, PREFIX_LEN>, KeyPrefix>,
    seen_leaf_nodes: Vec<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
}

A visitor of the radix tree which checks that the tree is well-formed.

In this context, well-formed means that in the tree:

there are no loops between nodes

every inner node has a number of children that is in range for the inner node type. For example, InnerNode16 has between 5 and 16 children.

the elements of the key (as part of inner node prefixes and child pointers) combine to match the leaf node key prefix

the previous and next pointers that form a doubly-linked list of leaf nodes has no loops, and the ordering of the leaves in the list is equal to the ordering of the leaves when sorted by key. The linked list should also be properly terminated with previous = None at the start and next = None at the end.

#1 and #3 are unlikely, but #2 is a possibility if specific tree operations are not implemented correctly. This visitor can be used to sanity check the tree in unit tests or other test cases.

This checker will only return a single issue at a time. A tree is only “well-formed” (by the definition given above) if the checker returns Ok(()).

Fields
current_key_prefix: Vec<u8>seen_nodes: HashMap<OpaqueNodePtr<K, V, PREFIX_LEN>, KeyPrefix>seen_leaf_nodes: Vec<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>
Implementations

impl<K, V, const PREFIX_LEN: usize> WellFormedChecker<K, V, PREFIX_LEN>
where K: AsBytes,

pub fn check<A: Allocator>( tree: &TreeMap<K, V, PREFIX_LEN, A>, ) -> Result<WellFormedTreeStats, MalformedTreeError<K, V, PREFIX_LEN>>

Traverse the given tree and check that it is well-formed. Returns the number of leaf nodes in the tree.

Errors

Returns an error if the given tree is not well-formed.

pub(crate) unsafe fn check_tree( tree: OpaqueNodePtr<K, V, PREFIX_LEN>, ) -> Result<WellFormedTreeStats, MalformedTreeError<K, V, PREFIX_LEN>>

Traverse the given tree and check that it is well-formed. Returns the number of leaf nodes in the tree.

Safety

For the duration of this function, the given node and all its children nodes must not be mutated.

Errors

Returns an error if the given tree is not well-formed.

fn verify_leaves_linked_list( &self, ) -> Result<(), MalformedTreeError<K, V, PREFIX_LEN>>

fn visit_inner_node<N>( &mut self, inner_node: &N, ) -> Result<WellFormedTreeStats, MalformedTreeError<K, V, PREFIX_LEN>>
where N: InnerNode<PREFIX_LEN, Key = K, Value = V>,