blart::raw::visitor

Trait Visitor

pub trait Visitor<K, V, const PREFIX_LEN: usize>: Sized {
    type Output;

    // Required methods
    fn default_output(&self) -> Self::Output;
    fn combine_output(&self, o1: Self::Output, o2: Self::Output) -> Self::Output;

    // Provided methods
    fn visit_inner_node<N>(&mut self, t: &N) -> Self::Output
       where N: InnerNode<PREFIX_LEN, Key = K, Value = V> + Visitable<K, V, PREFIX_LEN> { ... }
    fn visit_leaf(&mut self, t: &LeafNode<K, V, PREFIX_LEN>) -> Self::Output { ... }
}

The Visitor trait allows creating new operations on the radix tree by overriding specific handling methods for each of the node types.

Required Associated Types

type Output

The type of value that the visitor produces.

Required Methods

fn default_output(&self) -> Self::Output

Produce the default value of the Self::Output type.

fn combine_output(&self, o1: Self::Output, o2: Self::Output) -> Self::Output

Combine two instances of the Self::Output type for this Visitor.

Provided Methods

fn visit_inner_node<N>(&mut self, t: &N) -> Self::Output
where N: InnerNode<PREFIX_LEN, Key = K, Value = V> + Visitable<K, V, PREFIX_LEN>,

Visit an InnerNode.

fn visit_leaf(&mut self, t: &LeafNode<K, V, PREFIX_LEN>) -> Self::Output

Visit a LeafNode.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.