blart::raw::visitor
Module well_formed

Structs

KeyPrefix

A portion of an entire key that should uniquely identify each node in the tree.

WellFormedChecker

A visitor of the radix tree which checks that the tree is well-formed.

WellFormedTreeStats

This struct contains some simple stats collected from the trie when visiting it with WellFormedChecker.

Enums

MalformedTreeError

An issue with the well-formed-ness of the tree. See the documentation on WellFormedChecker for more context.