blart::raw::visitor::well_formed

Enum MalformedTreeError

#[non_exhaustive]
pub enum MalformedTreeError<K, V, const PREFIX_LEN: usize> {
    LoopFound {
        node_ptr: OpaqueNodePtr<K, V, PREFIX_LEN>,
        first_observed: KeyPrefix,
        later_observed: KeyPrefix,
    },
    WrongChildrenCount {
        key_prefix: KeyPrefix,
        inner_node_type: NodeType,
        num_children: usize,
    },
    PrefixMismatch {
        expected_prefix: KeyPrefix,
        entire_key: Vec<u8>,
    },
    EmptyTreeWithLen,
    LeafSiblingWrongOrder {
        sibling_key: Vec<u8>,
        leaf_key: Vec<u8>,
        expected_order: Ordering,
        actual_order: Ordering,
    },
    WrongSiblingLinks {
        leaf_key: Vec<u8>,
        expected_previous: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
        expected_next: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
        actual_previous: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
        actual_next: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>,
    },
}

An issue with the well-formed-ness of the tree. See the documentation on WellFormedChecker for more context.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

LoopFound

A loop was observed between nodes

Fields

node_ptr: OpaqueNodePtr<K, V, PREFIX_LEN>

The node that was observed more than once while traversing the tree

first_observed: KeyPrefix

The key prefix when the node was first observed

later_observed: KeyPrefix

The key prefix when the node was observed a second time

WrongChildrenCount

An inner node had an incorrect number of children

Fields

key_prefix: KeyPrefix

The key prefix identifying the inner node

inner_node_type: NodeType

The type of the inner node (InnerNode4, InnerNode16, etc)

This field is guaranteed not to be NodeType::Leaf

num_children: usize

The number of children found at the inner node

PrefixMismatch

The expected key prefix did not match the actual prefix that was present in the leaf

Fields

expected_prefix: KeyPrefix

The expected key prefix

entire_key: Vec<u8>

The entire key

EmptyTreeWithLen

The length of the tree is not 0, even though the root is Option::None

LeafSiblingWrongOrder

There is a leaf node with a sibling pointer where the sibling has the wrong key ordering with respect to the original leaf node.

Fields

sibling_key: Vec<u8>

The key bytes of the sibling leaf node

leaf_key: Vec<u8>

The key bytes of the leaf node that was out of order

expected_order: Ordering

The expected ordering (either Ordering::Greater or Ordering::Less) of the sibling and leaf keys.

actual_order: Ordering

The actual ordering (either Ordering::Greater or Ordering::Less) of the sibling and leaf keys.

WrongSiblingLinks

There is a leaf node which has incorrect values for either the previous or next sibling pointers.

Fields

leaf_key: Vec<u8>

The key bytes of the broken leaf node

expected_previous: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>

The expected previous pointer value

expected_next: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>

The expected next pointer value

actual_previous: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>

The actual previous pointer value

actual_next: Option<NodePtr<PREFIX_LEN, LeafNode<K, V, PREFIX_LEN>>>

The actual next pointer value