blart::raw::visitor::well_formed

Struct WellFormedTreeStats

pub struct WellFormedTreeStats {
    pub num_leaf: usize,
    pub num_inner: usize,
}

This struct contains some simple stats collected from the trie when visiting it with WellFormedChecker.

Fields
num_leaf: usize

The number of leaf nodes in the trie.

num_inner: usize

The number of inner node in the trie.

Implementations

impl WellFormedTreeStats

pub fn total_nodes(self) -> usize

The total number of leaf nodes and inner nodes in the trie.