Function blart::rust_nightly_apis::hasher_write_length_prefix

pub fn hasher_write_length_prefix<H: Hasher>(state: &mut H, num_entries: usize)

Writes a length prefix into this hasher, as part of being prefix-free.

If you’re implementing Hash for a custom collection, call this before writing its contents to this Hasher. That way (collection![1, 2, 3], collection![4, 5]) and (collection![1, 2], collection![3, 4, 5]) will provide different sequences of values to the Hasher

The impl<T> Hash for [T] includes a call to this method, so if you’re hashing a slice (or array or vector) via its Hash::hash method, you should not call this yourself.

This method is only for providing domain separation. If you want to hash a usize that represents part of the data, then it’s important that you pass it to Hasher::write_usize instead of to this method.

Note to Implementers

If you’ve decided that your Hasher is willing to be susceptible to Hash-DoS attacks, then you might consider skipping hashing some or all of the len provided in the name of increased performance.

This is a unstable API copied from the Rust standard library, tracking issue is #96762