blart::rust_nightly_apis

Function maybe_uninit_slice_assume_init_mut

pub unsafe fn maybe_uninit_slice_assume_init_mut<T>(
    slice: &mut [MaybeUninit<T>],
) -> &mut [T]

Assuming all the elements are initialized, get a mutable slice to them.

Safety

It is up to the caller to guarantee that the MaybeUninit<T> elements really are in an initialized state. Calling this when the content is not yet fully initialized causes undefined behavior.

See assume_init_mut for more details and examples.

This is a unstable API copied from the Rust standard library, tracking issue is #63569