blart::rust_nightly_apis
Macro likely

macro_rules! likely {
    ($b:expr) => { ... };
}

Hints to the compiler that branch condition is likely to be true. Returns the value passed to it.

Any use other than with if statements will probably not have an effect.

Note that, unlike most intrinsics, this is safe to call; it does not require an unsafe block. Therefore, implementations must not require the user to uphold any safety invariants.

This intrinsic does not have a stable counterpart.

This is a unstable API copied from the Rust standard library