blart

Struct ConcatTuple

pub struct ConcatTuple<M>(PhantomData<M>);

This type implements a BytesMapping for tuples of types, concatenating their byte representations together.

The M type parameter also takes a tuple, the same size as the input, which contains types also implementing BytesMapping. Each type in this tuple is used to transform the corresponding value in the input tuple.

Examples

Here is a basic example using the Identity transform for both tuple elements:

use blart::{ConcatTuple, AsBytes, Mapped, Identity};

let c1 = b"aaa";
let c2 = b"bb";

let t = Mapped::<ConcatTuple<(Identity, Identity)>, _>::new((*c1, *c2));

assert_eq!(t.as_bytes(), b"aaabb");

Here is a more complex example:

use std::net::Ipv4Addr;
use std::num::NonZeroI16;
use blart::{ConcatTuple, AsBytes, Mapped, ToOctets, ToIBE};

let c1 = NonZeroI16::new(256).unwrap();
let c2 = Ipv4Addr::LOCALHOST;

assert_eq!(Mapped::<ToIBE, _>::new(c1).as_bytes(), &[129, 0][..]);
assert_eq!(Mapped::<ToOctets, _>::new(c2).as_bytes(), &[127, 0, 0, 1][..]);

let t = Mapped::<ConcatTuple<(ToIBE, ToOctets)>, _>::new((c1, c2));

assert_eq!(t.as_bytes(), &[129, 0, 127, 0, 0, 1][..]);

Tuple Fields
0: PhantomData<M>