Struct blart::LeafNode

#[repr(align(8))]
pub struct LeafNode<K, V> {
    value: V,
    key: K,
}

Node that contains a single leaf value.

Fields
value: V

The leaf value.

key: K

The full key that the value was stored with.

Implementations

impl<K, V> LeafNode<K, V>

pub fn new(key: K, value: V) -> Self

Create a new leaf node with the given value.

pub fn key_ref(&self) -> &K

Returns a shared reference to the key contained by this leaf node

pub fn value_ref(&self) -> &V

Returns a shared reference to the value contained by this leaf node

pub fn value_mut(&mut self) -> &mut V

Returns a mutable reference to the value contained by this leaf node

pub fn entry_ref(&self) -> (&K, &V)

Return shared references to the key and value contained by this leaf node

pub fn entry_mut(&mut self) -> (&mut K, &mut V)

Return mutable references to the key and value contained by this leaf node

pub fn into_entry(self) -> (K, V)

Consume the leaf node and return a tuple of the key and value

pub fn matches_full_key<Q>(&self, possible_key: &Q) -> boolwhere K: Borrow<Q> + AsBytes, Q: AsBytes + ?Sized,

Check that the provided full key is the same one as the stored key.