Struct blart::Mapped

pub struct Mapped<B>where
    B: BytesMapping,{
    _mapping: PhantomData<B>,
    repr: B::Bytes,
}

A container for the bytestring that is produced from BytesMapping conversion

Fields
_mapping: PhantomData<B>repr: B::Bytes
Implementations

impl<B> Mapped<B>where B: BytesMapping,

pub fn new(value: B::Domain) -> Self

Transform a value into its ordered representation

pub fn get(self) -> B::Domain

Take the ordered representation and convert it back to the original value