Struct blart::tagged_pointer::TaggedPointer

#[repr(transparent)]
pub struct TaggedPointer<P, const MIN_BITS: u32>(NonNull<P>);

A non-null pointer type which carries several bits of metadata.

The MIN_BITS constant is used to ensure that any type that is used with this pointer will have sufficient alignment to carry the specified number of bits. As a result, we can cast safely because we know that the const-evaluation process & type-checking will ensure that the number of bits is sufficient.

Examples

Here is an example of successfully tagging a pointer and retrieving the data:

use blart::tagged_pointer::TaggedPointer;

let pointee = "Hello world!";
let pointer = Box::into_raw(Box::new(pointee));
let tag_data = 0b101usize;

let mut tagged_pointer = TaggedPointer::<&str, 3>::new_with_data(pointer, tag_data).unwrap();

assert_eq!(unsafe { *tagged_pointer.to_ptr() }, "Hello world!");
assert_eq!(tagged_pointer.to_data(), 0b101);

tagged_pointer.set_data(0b010);

assert_eq!(unsafe { *tagged_pointer.to_ptr() }, "Hello world!");
assert_eq!(tagged_pointer.to_data(), 0b010);

// Collecting the data into `Box` to safely drop it
unsafe {
    drop(Box::from_raw(tagged_pointer.to_ptr()));
}
Run

Here is an example of tagging a pointer, then casting it to a different type and retrieving the same data:

use blart::tagged_pointer::TaggedPointer;

let pointee = u64::MAX;
let pointer = Box::into_raw(Box::new(pointee));
let tag_data = 0b010usize;

let mut tagged_pointer = TaggedPointer::<u64, 3>::new_with_data(pointer, tag_data).unwrap();

assert_eq!(unsafe { *tagged_pointer.to_ptr() }, u64::MAX);
assert_eq!(tagged_pointer.to_data(), 0b010);

tagged_pointer.set_data(0b101);

let new_tagged_pointer = tagged_pointer.cast::<i64>();

assert_eq!(unsafe { *new_tagged_pointer.to_ptr() }, -1);
assert_eq!(new_tagged_pointer.to_data(), 0b101);

unsafe {
    drop(Box::from_raw(tagged_pointer.to_ptr()));
}
Run

Here is an example of a cast that fails to compile because the result type has insufficient alignment:

use blart::tagged_pointer::TaggedPointer;

let pointee = u64::MAX;
let pointer = Box::into_raw(Box::new(pointee));
let tag_data = 0b010usize;

let mut tagged_pointer = TaggedPointer::<u64, 3>::new_with_data(pointer, tag_data).unwrap();

assert_eq!(unsafe { *tagged_pointer.to_ptr() }, u64::MAX);
assert_eq!(tagged_pointer.to_data(), 0b010);

tagged_pointer.set_data(0b101);

// This cast to `i32` causes the compilation failure
let new_tagged_pointer = tagged_pointer.cast::<i32>();
Run

Tuple Fields
0: NonNull<P>
Implementations

impl<P, const MIN_BITS: u32> TaggedPointer<P, MIN_BITS>

pub const ALIGNMENT: usize = _

The ABI-required minimum alignment of the P type.

The maximum alignment of a rust type is 2^29 according to the reference. On 32-bit and 64-bit platforms that means that it is impossible for the POINTER_MASK to be zero, there must be at least 3 bit present.

pub const DATA_MASK: usize = _

A mask for data-carrying bits of the pointer.

pub const NUM_BITS: u32 = _

Number of available bits of storage in the pointer.

pub const POINTER_MASK: usize = _

A mask for the non-data-carrying bits of the pointer.

pub fn new(pointer: *mut P) -> Option<TaggedPointer<P, MIN_BITS>>

Create a new tagged pointer from a possibly null pointer.

If the pointer is null, returns None. Otherwise, returns a tagged pointer.

Panics

Panics if the given pointer is not aligned according to the minimum alignment required for the P type.

pub unsafe fn new_unchecked(pointer: *mut P) -> TaggedPointer<P, MIN_BITS>

Create a new non-null tagged pointer without verifying that it is non-null.

Panics

Panics if the given pointer is not aligned according to the minimum alignment required for the P type.

Safety

The pointer value must not be null.

pub fn new_with_data( pointer: *mut P, data: usize, ) -> Option<TaggedPointer<P, MIN_BITS>>

Create a new tagged pointer and immediately set data on it.

Returns None if the given pointer is null.

Panics

Panics if the given pointer is not aligned according to the minimum alignment required for the P type.

pub fn to_ptr(self) -> *mut P

Consume this tagged pointer and produce a raw mutable pointer to the memory location.

This pointer is guaranteed to be non-null.

pub fn to_data(self) -> usize

Consume this tagged pointer and produce the data it carries.

pub fn set_data(&mut self, data: usize)

Update the data this tagged pointer carries to a new value.

Panics

Panics if any bits other than the lowest Self::NUM_BITS are non-zero in the new data value.

pub fn cast<Q>(self) -> TaggedPointer<Q, MIN_BITS>

Casts to a TaggedPointer of another type.

This function will transfer the data bits from the original pointer to the new pointer.

Errors

This function will error if the alignment of the new type does not equal the alignment of the existing type. This is because the number of data-carrying bits could be different.