blart::testing

Function generate_key_fixed_length

pub fn generate_key_fixed_length<const KEY_LENGTH: usize>(
    level_widths: [u8; KEY_LENGTH],
) -> impl Iterator<Item = [u8; KEY_LENGTH]>

Generate an iterator of bytestring keys, all with the same length.

The level_widths argument specifies the number of values generated per digit of the array. For example, using [3, 2, 1] will generate keys of length 3. The generate keys will have 4 (3 + 1) unique values for the first digit, 3 unique values for the second digit, and 2 unique values for the last digit. In general, this iterator will produce (level_widths[0] + 1) * (level_widths[1] + 1) * ... * (level_widths[KEY_LENGTH - 1] + 1) keys in total.

Examples

let keys = generate_key_fixed_length([3, 2, 1]).collect::<Vec<_>>();
assert_eq!(keys.len(), 24);
assert_eq!(keys[0].as_ref(), &[0, 0, 0]);
assert_eq!(keys[keys.len() / 2].as_ref(), &[2, 0, 0]);
assert_eq!(keys[keys.len() - 1].as_ref(), &[3, 2, 1]);

for k in keys {
    println!("{:?}", k);
}

The above example will print

[0, 0, 0]
[0, 0, 1]
[0, 1, 0]
[0, 1, 1]
[0, 2, 0]
[0, 2, 1]
[1, 0, 0]
[1, 0, 1]
[1, 1, 0]
[1, 1, 1]
[1, 2, 0]
[1, 2, 1]
[2, 0, 0]
[2, 0, 1]
[2, 1, 0]
[2, 1, 1]
[2, 2, 0]
[2, 2, 1]
[3, 0, 0]
[3, 0, 1]
[3, 1, 0]
[3, 1, 1]
[3, 2, 0]
[3, 2, 1]

Panics

Panics if max_len is 0.

Panics if value_stops is 0.