blart::testing

Function generate_key_with_prefix

pub fn generate_key_with_prefix<const KEY_LENGTH: usize>(
    level_widths: [u8; KEY_LENGTH],
    prefix_expansions: impl AsRef<[PrefixExpansion]>,
) -> impl Iterator<Item = Box<[u8]>>

Generate an iterator of fixed length bytestring keys, where specific portions of the key are expanded as duplicate bytes.

This is meant to simulate keys with shared prefixes in different portions of the key string.

Examples

let keys = generate_key_with_prefix([2; 3], [PrefixExpansion { base_index: 0, expanded_length: 3 }]).collect::<Vec<_>>();
assert_eq!(keys.len(), 27);
assert_eq!(keys[0].as_ref(), &[0, 0, 0, 0, 0]);
assert_eq!(keys[(keys.len() / 2) - 2].as_ref(), &[1, 1, 1, 0, 2]);
assert_eq!(keys[keys.len() - 1].as_ref(), &[2, 2, 2, 2, 2]);

for k in keys {
    println!("{:?}", k);
}

The above example will print out:

[0, 0, 0, 0, 0]
[0, 0, 0, 0, 1]
[0, 0, 0, 0, 2]
[0, 0, 0, 1, 0]
[0, 0, 0, 1, 1]
[0, 0, 0, 1, 2]
[0, 0, 0, 2, 0]
[0, 0, 0, 2, 1]
[0, 0, 0, 2, 2]
[1, 1, 1, 0, 0]
[1, 1, 1, 0, 1]
[1, 1, 1, 0, 2]
[1, 1, 1, 1, 0]
[1, 1, 1, 1, 1]
[1, 1, 1, 1, 2]
[1, 1, 1, 2, 0]
[1, 1, 1, 2, 1]
[1, 1, 1, 2, 2]
[2, 2, 2, 0, 0]
[2, 2, 2, 0, 1]
[2, 2, 2, 0, 2]
[2, 2, 2, 1, 0]
[2, 2, 2, 1, 1]
[2, 2, 2, 1, 2]
[2, 2, 2, 2, 0]
[2, 2, 2, 2, 1]
[2, 2, 2, 2, 2]

Panics

Panics if base_key_len is 0.

Panics if value_stops is 0.

Panics if any PrefixExpansion has expanded_length equal to 0.

Panics if any PrefixExpansion has base_index greater than or equal to base_key_len.