blart::testing

Function generate_keys_skewed

pub fn generate_keys_skewed(max_len: usize) -> impl Iterator<Item = Box<[u8]>>

Generate an iterator of bytestring keys, with increasing length up to a maximum value.

This iterator will produce max_len number of keys. Each key has the form [0*, u8::MAX], meaning zero or more 0 values, followed by a single u8::MAX value. The final u8::MAX value is added to ensure that no key is a prefix of another key generated by this function.

Examples

let keys = generate_keys_skewed(10).collect::<Vec<_>>();
assert_eq!(keys.len(), 10);
assert_eq!(keys[0].as_ref(), &[255]);
assert_eq!(keys[keys.len() - 1].as_ref(), &[0, 0, 0, 0, 0, 0, 0, 0, 0, 255]);

for k in keys {
    println!("{:?}", k);
}

The above example will print

[255]
[0, 255]
[0, 0, 255]
[0, 0, 0, 255]
[0, 0, 0, 0, 255]
[0, 0, 0, 0, 0, 255]
[0, 0, 0, 0, 0, 0, 255]
[0, 0, 0, 0, 0, 0, 0, 255]
[0, 0, 0, 0, 0, 0, 0, 0, 255]
[0, 0, 0, 0, 0, 0, 0, 0, 0, 255]

Panics

Panics if max_len is 0.