Trait blart::InnerNode

pub trait InnerNode: Node {
    type GrownNode: InnerNode<Value = <Self as Node>::Value>;
    type Iter: Iterator<Item = (u8, OpaqueNodePtr<<Self as Node>::Value>)> + Into<InnerNodeIter<<Self as Node>::Value>>;

    fn lookup_child(
        &self, 
        key_fragment: u8
    ) -> Option<OpaqueNodePtr<Self::Value>>;
    fn write_child(
        &mut self, 
        key_fragment: u8, 
        child_pointer: OpaqueNodePtr<Self::Value>
    );
    fn grow(&self) -> Self::GrownNode;
    fn header(&self) -> &Header;
    fn header_mut(&mut self) -> &mut Header;
    unsafe fn iter(&self) -> Self::Iter;

    fn is_full(&self) -> bool { ... }
}

Common methods implemented by all inner node.

Required Associated Types

type GrownNode: InnerNode<Value = <Self as Node>::Value>

The type of the next larger node type.

type Iter: Iterator<Item = (u8, OpaqueNodePtr<<Self as Node>::Value>)> + Into<InnerNodeIter<<Self as Node>::Value>>

The type of the iterator over all children of the inner node

Required Methods

fn lookup_child(&self, key_fragment: u8) -> Option<OpaqueNodePtr<Self::Value>>

Search through this node for a child node that corresponds to the given key fragment.

fn write_child(
&mut self,
key_fragment: u8,
child_pointer: OpaqueNodePtr<Self::Value>
)

Write a child pointer with key fragment to this inner node.

If the key fragment already exists in the node, overwrite the existing child pointer.

Panics

Panics when the node is full.

fn grow(&self) -> Self::GrownNode

Grow this node into the next larger class, copying over children and prefix information.

fn header(&self) -> &Header

Access the header information for this node.

fn header_mut(&mut self) -> &mut Header

Access the header information for this node.

unsafe fn iter(&self) -> Self::Iter

Create an iterator over all (key bytes, child pointers) in this inner node.

Safety

The iterator type does not carry any lifetime, so the caller of this function must enforce that the lifetime of the iterator does not overlap with any mutating operations on the node.

Provided Methods

fn is_full(&self) -> bool

Returns true if this node has no more space to store children.