Trait blart::OrderedBytes

pub unsafe trait OrderedBytes: AsBytes + Ord { }

This trait is used to mark types where the lexicographic ordering of their byte representation (as output by AsBytes::as_bytes) matches their normal ordering (as determined by Ord).

Safety

This trait can only be implemented if the above condition holds.

Object Safety

This trait is not object safe.