blart
Module visitor

Utilities for inspecting the trie structure.

Structs

DebugAsDisplay

This struct is intended to wrap a type which implements Debug and expose it as an implementation of Display.

DotPrinter

A visitor of the radix trie that will print the tree in “dot” notation.

DotPrinterSettings

Settings which customize the output of the DotPrinter visitor.

InnerNodeStats

Statistics for inner nodes

KeyPrefix

A portion of an entire key that should uniquely identify each node in the tree.

LeafStats

Statistics for inner nodes

TreeStats

Collection of stats about the number of nodes types present in a tree

TreeStatsCollector

A visitor of the radix tree which collects statistics about the tree, like how many inner nodes of each type, how many leaves

WellFormedChecker

A visitor of the radix tree which checks that the tree is well-formed.

WellFormedTreeStats

This struct contains some simple stats collected from the trie when visiting it with WellFormedChecker.

Enums

MalformedTreeError

An issue with the well-formed-ness of the tree. See the documentation on WellFormedChecker for more context.

Traits

Visitable

The Visitable trait allows Visitors to traverse the structure of the implementing type and produce some output.

Visitor

The Visitor trait allows creating new operations on the radix tree by overriding specific handling methods for each of the node types.