Struct blart::visitor::DotPrinter

pub struct DotPrinter<O: Write> {
    output: O,
    next_id: usize,
    settings: DotPrinterSettings,
}

A visitor of the radix trie that will print the tree in “dot” notation.

See ‘DOT Language | Graphviz’ for information about syntax and example of the language.

Fields
output: Onext_id: usizesettings: DotPrinterSettings
Implementations

impl<O: Write> DotPrinter<O>

pub fn print<K, V, const PREFIX_LEN: usize>( output: O, tree: &TreeMap<K, V, PREFIX_LEN>, settings: DotPrinterSettings, ) -> Option<Result<()>>
where K: Display, V: Display,

Write the dot-format of the given tree to the given output.

unsafe fn print_tree<K, V, const PREFIX_LEN: usize>( output: O, tree: &OpaqueNodePtr<K, V, PREFIX_LEN>, settings: DotPrinterSettings, ) -> Result<()>
where K: Display, V: Display,

Write the dot-format of the given tree to the given output.

Safety

For the duration of this function, the given node and all its children nodes must not get mutated.

fn output_prelude(&mut self) -> Result<()>

fn output_epilogue(&mut self) -> Result<()>

fn get_id(&mut self) -> usize

fn write_inner_node<K, T, N, const PREFIX_LEN: usize>( &mut self, inner_node: &N, ) -> Result<usize>
where K: Display, T: Display, N: InnerNode<PREFIX_LEN, Key = K, Value = T>,