Trait blart::visitor::Visitable

pub trait Visitable<K, T> {
    // Required method
    fn super_visit_with<V: Visitor<K, T>>(&self, visitor: &mut V) -> V::Output;

    // Provided method
    fn visit_with<V: Visitor<K, T>>(&self, visitor: &mut V) -> V::Output { ... }
}

The Visitable trait allows Visitors to traverse the structure of the implementing type and produce some output.

Required Methods

fn super_visit_with<V: Visitor<K, T>>(&self, visitor: &mut V) -> V::Output

This function provides the default traversal behavior for the implementing type.

The implementation should call visit_with(visitor) for all relevant sub-fields of the type. If there are no relevant sub-fields, it should just produce the default output.

Provided Methods

fn visit_with<V: Visitor<K, T>>(&self, visitor: &mut V) -> V::Output

This function will traverse the implementing type and execute any specific logic from the given Visitor.

This function should be override for types that have corresponding hooks in the Visitor trait. For example the Visitable implementation for InnerNode4 looks like:

impl<K, T> Visitable<K, T> for InnerNode4<K, T> {
    ...
  
    fn visit_with<V: Visitor<K, T>>(&self, visitor: &mut V) -> V::Output {
        visitor.visit_node4(self)
    }
}
Run

The call to visitor.visit_node4(self) allows the visitor to execute specific handling logic.