Crate blart

Adaptive radix trie implementation

References

Leis, V., Kemper, A., & Neumann, T. (2013, April). The adaptive radix tree: ARTful indexing for main-memory databases. In 2013 IEEE 29th International Conference on Data Engineering (ICDE) (pp. 38-49). IEEE. Link to PDF

Re-exports

pub use raw::visitor;

Modules

alloc 🔒

bytes 🔒

collections 🔒

map

Module containing implementations of the TreeMap and associated iterators/etc.

raw

Trie node representation and manipulation

rust_nightly_apis 🔒

Module containing copies of Rust standard library unstable functions for use outside of the nightly distribution.

tagged_pointer 🔒

A tagged pointer is a pointer (concretely a memory address) with additional data associated with it, such as an indirection bit or reference count. This additional data is often “folded” into the pointer, meaning stored inline in the data representing the address, taking advantage of certain properties of memory addressing.

Structs

ConcatTuple

This type implements a BytesMapping for tuples of types, concatenating their byte representations together.

Identity

This type implements a BytesMapping that preserves the original type without converting it to bytes.

Mapped

A container for the bytestring that is produced from BytesMapping conversion

ToIBE

This struct represents a conversion of signed integers to a format that allows the natural ordering of the numbers to match the lexicographic ordering of the bytes.

ToOctets

This struct represents a conversion of IP addresses (V4 and V6) into their component bytes.

ToUBE

This struct represents a conversion of unsigned integers to the big endian format, so that the natural ordering of the numbers matches the lexicographic ordering of the bytes.

TreeMap

An ordered map based on an adaptive radix tree.

Traits

AsBytes

Any type implementing AsBytes can be decomposed into bytes.

BytesMapping

Trait representing a reversible conversion from a type to some sort of byte string.

NoPrefixesBytes

This trait is used to mark types which have a byte representation which is guaranteed to not be a prefix of any other value of the same type.

OrderedBytes

This trait is used to mark types where the lexicographic ordering of their byte representation (as output by AsBytes::as_bytes) matches their normal ordering (as determined by Ord).