Crate blart

Adaptive radix trie implementation

References

Leis, V., Kemper, A., & Neumann, T. (2013, April). The adaptive radix tree: ARTful indexing for main-memory databases. In 2013 IEEE 29th International Conference on Data Engineering (ICDE) (pp. 38-49). IEEE. Link to PDF

Modules

tagged_pointer

A tagged pointer is a pointer (concretely a memory address) with additional data associated with it, such as an indirection bit or reference count. This additional data is often “folded” into the pointer, meaning stored inline in the data representing the address, taking advantage of certain properties of memory addressing.

visitor

Utilities for inspecting the trie structure.

Structs

Header

The common header for all inner nodes

InnerNode48

Node that references between 17 and 49 children

InnerNode48Iter

An iterator over all the children of a InnerNode48.

InnerNode256

Node that references between 49 and 256 children

InnerNode256Iter

An iterator over all the children of a InnerNode256.

InnerNodeCompressed

Node type that has a compact representation for key bytes and children pointers.

InnerNodeCompressedIter

An iterator all the children of an InnerNodeCompressed.

LeafNode

Node that contains a single leaf value.

NodePtr

A pointer to a Node{4,16,48,256}.

OpaqueNodePtr

An opaque pointer to a Node{4,16,48,256}.

RestrictedNodeIndex

A restricted index only valid from 0 to LIMIT - 1.

TrieRangeFullIter

An iterator over all the entries in a tree.

TryFromByteError

The error type returned when attempting to construct an index outside the accepted range of a PartialNodeIndex.

Enums

ConcreteNode

An enum the encapsulates every type of Node

ConcreteNodePtr

An enum that encapsulates pointers to every type of Node

InnerNodeIter

A generic iterator that uses specific iterators for each NodeType (excluding leaves) inside.

InsertError

The error type for the insert operation on the tree.

NodeType

The representation of inner nodes

Constants

NUM_PREFIX_BYTES

The number of bytes stored for path compression in the node header.

Traits

InnerNode

Common methods implemented by all inner node.

Node

All nodes which contain a runtime tag that validates their type.

Functions

deallocate_tree⚠

Deallocate the given node and all children of the given node.

insert_unchecked⚠

Insert the given key-value pair into the tree.

maximum_unchecked⚠

Search for the leaf with the maximum key, by lexicographic ordering.

minimum_unchecked⚠

Search for the leaf with the minimum key, by lexicographic ordering.

search_unchecked⚠

Search in the given tree for the value stored with the given key.

Type Definitions

InnerNode4

Node that references between 2 and 4 children

InnerNode16

Node that references between 5 and 16 children