Module blart::bytes::mapped

Macros

as_bytes_for_tuples 🔒

impl_ord_for_mapped 🔒

impl_ordered_bytes_ints 🔒

impl_ordered_bytes_nonzero_ints 🔒

sum 🔒

Structs

Concat

This type implements a BytesMapping for tuples of types, concatenating their byte representations together.

Identity

This type implements a BytesMapping that preserves the original type without converting it to bytes.

Mapped

A container for the bytestring that is produced from BytesMapping conversion

ToIBE

This struct represents a conversion of signed integers to a format that allows the natural ordering of the numbers to match the lexicographic ordering of the bytes.

ToOctets

This struct represents a conversion of IP addresses (V4 and V6) into their component bytes. The ordering of IP addresses is already the lexicographic ordering of the component bytes, so it will be preserved.

ToUBE

This struct represents a conversion of unsigned integers to the big endian format, so that the natural ordering of the numbers matches the lexicographic ordering of the bytes.

Traits

BytesMapping

Trait representing a reversible conversion from a type to some sort of byte string.