blart::bytes::mapped

Macro impl_ordered_bytes_ints_arrays

macro_rules! impl_ordered_bytes_ints_arrays {
    ($([$mapping:ty => $($elem:ty),+]),*) => { ... };
    ($mapping:ty;$elem:ty;array) => { ... };
    ($mapping:ty;$elem:ty;vec_like;$domain:ty) => { ... };
}

Safety

This macro must only be used with types $elem where size_of::<$elem> == <$elem as AsBytes>::as_bytes(&instance).len(). The types must also be fixed size, so all instance of $elem must have the same size.

This macro must only be used with types that implement

impl BytesMapping<$elem> for $mapping {
    // ...
}

$elem must also implement

unsafe impl OrderedBytes for Mapped<$mapping, $elem> {}