blart::collections::map::iterators

Module range

This module contains types and functions relating to iterating over a range of the TreeMap.

Macros

implement_range_iter 🔒

Structs

InnerNodeSearchResult 🔒

This struct contains details of where and why the search stopped in an inner node.

Range

An iterator over a sub-range of entries in a TreeMap.

RangeMut

A mutable iterator over a sub-range of entries in a TreeMap.

Enums

InnerNodeSearchResultReason 🔒

These are search termination reasons specific to inner nodes.

TerminatingNodeSearchResult 🔒

This enum provides the different cases of search result when looking for a starting/ending point for iteration.

Functions

find_leaf_pointer_for_bound 🔒 ⚠

This function searches a trie for the smallest/largest leaf node that is greater/less than the given bound.

find_terminating_node 🔒 ⚠

Find the node (inner or leaf) that is identified by the given bytestring.

validate_range_bounds 🔒

Panic if the given range bounds

Type Aliases

KeyByteAndChild 🔒