blart::collections::map::iterators

Module subtree_iter

Macros

implement_prefix_iter 🔒

Structs

SubtreeIter

An iterator over a subtree in a TreeMap.

SubtreeIterMut

A mutable iterator over a range of entries that all have the same key prefix in a TreeMap.