Module blart::map

Module containing implementations of the TreeMap and associated iterators/etc.

Structs

Fuzzy

An iterator over all the LeafNodes within a specific edit distance

FuzzyMut

An iterator over all the LeafNodes within a specific edit distance

IntoIter

An owning iterator over the entries of a TreeMap.

IntoKeys

An owning iterator over the keys of a TreeMap.

IntoValues

An owning iterator over the values of a TreeMap.

Iter

An iterator over all the LeafNodes

IterMut

An iterator over all the LeafNodes

Keys

An iterator over all the LeafNodes

OccupiedEntry

A view into an occupied entry in a TreeMap. It is part of the Entry enum.

OccupiedEntryRef

A view into an occupied entry in a TreeMap. It is part of the EntryRef enum.

Prefix

An iterator over a range of entries that all have the same key prefix in a TreeMap.

PrefixMut

A mutable iterator over a range of entries that all have the same key prefix in a TreeMap.

Range

An iterator over a sub-range of entries in a TreeMap.

RangeMut

A mutable iterator over a sub-range of entries in a TreeMap.

TreeMap

An ordered map based on an adaptive radix tree.

VacantEntry

A view into a vacant entry in a TreeMap. It is part of the Entry enum.

VacantEntryRef

A view into a vacant entry in a TreeMap. It is part of the EntryRef enum.

Values

An iterator over all the LeafNodes

ValuesMut

An iterator over all the LeafNodes

Enums

Entry

A view into a single entry in a map, which may either be vacant or occupied.

EntryRef

A view into a single entry in a map, which may either be vacant or occupied.