Module blart::nodes::operations::delete

Structs

DeleteResult

The results of a successful delete operation

DeleteSearchResult 🔒

Functions

delete_maximum_unchecked⚠

Find and delete the maximum leaf in the tree, returning the maximum LeafNode.

delete_minimum_unchecked⚠

Find and delete the minimum leaf in the tree, returning the minimum LeafNode.

delete_unchecked⚠

Removes a key from the tree, returning the LeafNode corresponding to the key if the key was previously in the tree.

find_maximum_to_delete 🔒 ⚠

Find the maximum leaf in the tree and return the information necessary to delete it.

find_minimum_to_delete 🔒 ⚠

Find the minimum leaf in the tree and return the information necessary to delete it.

inner_delete_non_root_unchecked 🔒 ⚠

Delete the given non-root leaf node.

inner_delete_unchecked 🔒 ⚠

Handle the logic of deleting a leaf node from the tree, after it has been found.

remove_child_from_inner_node_and_compress 🔒 ⚠

Remove a child node from the given inner node, return the child node pointer if it was compressed.

search_for_node_to_delete 🔒 ⚠

Search in the given tree for the leaf node to delete, returning None if it does not exist.