Function blart::nodes::operations::lookup::check_prefix_lookup_child

pub(crate) unsafe fn check_prefix_lookup_child<K, V, N, const PREFIX_LEN: usize>(
    inner_ptr: NodePtr<PREFIX_LEN, N>,
    key_bytes: &[u8],
    current_depth: &mut usize,
    prefix_match_behavior: &mut PrefixMatchBehavior,
) -> Option<OpaqueNodePtr<K, V, PREFIX_LEN>>
where
    N: InnerNode<PREFIX_LEN, Key = K, Value = V>,
    K: AsBytes,

For the given InnerNode, check the node prefix, then lookup the child based on the search depth.

If the prefix does not match, it returns None. If there is no matching child for the key byte, it returns None.

Safety

No other access or mutation to the inner_ptr Node can happen while this function runs.