Module blart::nodes::representation

Trie node representation

Re-exports

pub use inner_node_256::*;

pub use inner_node_48::*;

pub use inner_node_compressed::*;

Modules

header 🔒

Different header type

inner_node_48 🔒

inner_node_256 🔒

inner_node_compressed 🔒

private 🔒

Structs

LeafNode

Node that contains a single leaf value.

Mismatch

Represents a prefix mismatch

NodePtr

A pointer to a Node.

OpaqueNodePtr

An opaque pointer to a Node.

OpaqueValue 🔒

A placeholder type that has the required amount of alignment.

Enums

ConcreteNodePtr

An enum that encapsulates pointers to every type of Node

MatchPrefixResult

Result of prefix match

NodeType

The representation of inner nodes

Traits

InnerNode

Common methods implemented by all inner node.

Node

All nodes which contain a runtime tag that validates their type.