blart::raw::operations::clone

Function clone_unchecked

pub unsafe fn clone_unchecked<K: Clone + AsBytes, V: Clone, A: Allocator, const PREFIX_LEN: usize>(
    root: OpaqueNodePtr<K, V, PREFIX_LEN>,
    alloc: &A,
) -> CloneResult<K, V, PREFIX_LEN>

Clone the given trie and all the key-values paired contained.

This function does not use recursion to clone, so it should not cause stack overflow when cloning a deep trie.

Safety

This function can only be called while there are no mutable references to any of the nodes in this trie.