blart::raw::operations

Module deallocate

Functions

deallocate_inner_node 🔒 ⚠

This function will read an InnerNode and place all children that satisfy the filter_children predicate on the stack. Then it will deallocate the given InnerNode.

deallocate_leaves⚠

Deallocate all the leaf nodes in the linked list starting that are within the given iterator.

deallocate_tree⚠

Deallocate the given node and all children of the given node.

deallocate_tree_non_leaves⚠

Deallocate the given node and all children of the given node except for leaf nodes.