blart::raw::operations::deallocate

Function deallocate_tree_non_leaves

pub unsafe fn deallocate_tree_non_leaves<K, V, const PREFIX_LEN: usize, A: Allocator>(
    root: OpaqueNodePtr<K, V, PREFIX_LEN>,
    alloc: &A,
)

Deallocate the given node and all children of the given node except for leaf nodes.

Safety

This function must only be called once for this root node and all descendants, otherwise a double-free could result.

This function should not be called concurrently with any read of the tree, otherwise it could result in a use-after-free.

alloc must be the same allocator which was used to allocate all the inner nodes of the tree.