blart::raw::operations

Module delete

Structs

DeletePoint

This struct represents a location in the trie that can be deleted.

DeleteResult

The results of a successful delete operation

Functions

find_maximum_to_delete⚠

Find the maximum leaf in the tree and return the information necessary to delete it.

find_minimum_to_delete⚠

Find the minimum leaf in the tree and return the information necessary to delete it.

inner_delete_non_root_unchecked 🔒 ⚠

Delete the given non-root leaf node.

remove_child_from_inner_node_and_compress 🔒 ⚠

Remove a child node from the given inner node, return the child node pointer if it was compressed.

search_for_delete_point⚠

Search in the given tree for the leaf node to delete, returning None if it does not exist.