blart::raw

Module representation

Trie node representation

Re-exports

pub use inner_node_256::*;

pub use inner_node_48::*;

pub use inner_node_compressed::*;

Modules

header 🔒

Different header type

inner_node_48 🔒

inner_node_256 🔒

inner_node_compressed 🔒

private 🔒

Macros

concrete_inner_node_ptr_from 🔒

concrete_node_ptr_from 🔒

Structs

AttemptOptimisticPrefixMatch

This struct represents a successful match against a prefix using the InnerNode::attempt_pessimistic_match_prefix function.

ExplicitMismatch

Represents a prefix mismatch when looking at the entire prefix, including in cases where it is read from a child leaf node.

LeafNode

Node that contains a single leaf value.

NodePtr

A pointer to a Node.

OpaqueNodePtr

An opaque pointer to a Node.

OpaqueValue 🔒

A placeholder type that has the required amount of alignment.

OptimisticMismatch

Represents a prefix mismatch when looking only at the prefix content present in an InnerNode header.

PessimisticMismatch

Represents a prefix mismatch when looking only at the prefix content present in an InnerNode header.

PrefixMatch

This struct represents a successful match against a prefix using either the InnerNode::optimistic_match_prefix or InnerNode::match_full_prefix functions.

TreePathSearch

This type is used to track the parent and grandparent when searching down the tree.

Enums

ConcreteInnerNodePtr

An enum that encapsulates pointers to every type of InnerNode

ConcreteNodePtr

An enum that encapsulates pointers to every type of Node

NodeType

The representation of inner nodes

TreePath

This enum represents different kinds of tree paths pointing to a leaf node.

Traits

InnerNode

Common methods implemented by all inner node.

Node

All nodes which contain a runtime tag that validates their type.

Type Aliases

OptionalLeafPtr 🔒

This type alias represents an optional pointer to a leaf node.