blart

Module rust_nightly_apis

Module containing copies of Rust standard library unstable functions for use outside of the nightly distribution.

Modules

ptr 🔒

This module contains shim functions for strict-provenance stuff related to pointers

Macros

assume 🔒

Informs the optimizer that a condition is always true. If the condition is false, the behavior is undefined.

likely 🔒

Hints to the compiler that branch condition is likely to be true. Returns the value passed to it.

unlikely 🔒

Hints to the compiler that branch condition is likely to be false. Returns the value passed to it.

Functions

box_new_uninit_slice

Constructs a new boxed slice with uninitialized contents.

hasher_write_length_prefix

Writes a length prefix into this hasher, as part of being prefix-free.

maybe_uninit_slice_assume_init_mut⚠

Assuming all the elements are initialized, get a mutable slice to them.

maybe_uninit_slice_assume_init_ref⚠

Assuming all the elements are initialized, get a slice to them.