blart

Module rust_nightly_apis

Module containing copies of Rust standard library unstable functions for use outside of the nightly distribution.

Modules

ptr 🔒

This module contains shim functions for strict-provenance stuff related to pointers

Macros

likely 🔒

Hints to the compiler that a branch condition is unlikely to be true. Returns the value passed to it.

unlikely 🔒

Hints to the compiler that given path is cold, i.e., unlikely to be taken. The compiler may choose to optimize paths that are not cold at the expense of paths that are cold.

Functions

hasher_write_length_prefix

Writes a length prefix into this hasher, as part of being prefix-free.

maybe_uninit_slice_assume_init_mut⚠

Assuming all the elements are initialized, get a mutable slice to them.

maybe_uninit_slice_assume_init_ref⚠

Assuming all the elements are initialized, get a slice to them.