blart

Struct Mapped

#[repr(transparent)]
pub struct Mapped<B, D>
where
    B: BytesMapping<D>,
{
    _mapping: PhantomData<(B, D)>,
    repr: B::Bytes,
}

A container for the bytestring that is produced from BytesMapping conversion

Fields
_mapping: PhantomData<(B, D)>repr: B::Bytes
Implementations

impl<B, D> Mapped<B, D>
where B: BytesMapping<D>,

pub fn new(value: D) -> Self

Transform a value into its bytes representation

fn with_repr(repr: B::Bytes) -> Self

Created a new instance of Mapped starting from the byte representation

pub fn get(self) -> D

Take the byte representation and convert it back to the original value