blart

Trait AsBytes

pub trait AsBytes {
    // Required method
    fn as_bytes(&self) -> &[u8] ;
}

Any type implementing AsBytes can be decomposed into bytes.

The primary purpose of this trait is to allow different types to be used as keys on the crate::TreeMap and TreeSet types.

Required Methods

fn as_bytes(&self) -> &[u8]

View the current value as a byte array.