blart

Trait OrderedBytes

pub unsafe trait OrderedBytes: AsBytes + Ord { }

This trait is used to mark types where the lexicographic ordering of their byte representation (as output by AsBytes::as_bytes) matches their normal ordering (as determined by Ord).

Safety

This trait can only be implemented if the above condition holds.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.