Trait blart::visitor::Visitor

pub trait Visitor<K, V>: Sized {
    type Output;

    // Required methods
    fn default_output(&self) -> Self::Output;
    fn combine_output(&self, o1: Self::Output, o2: Self::Output) -> Self::Output;

    // Provided methods
    fn visit_node4(&mut self, t: &InnerNode4<K, V>) -> Self::Output { ... }
    fn visit_node16(&mut self, t: &InnerNode16<K, V>) -> Self::Output { ... }
    fn visit_node48(&mut self, t: &InnerNode48<K, V>) -> Self::Output { ... }
    fn visit_node256(&mut self, t: &InnerNode256<K, V>) -> Self::Output { ... }
    fn visit_leaf(&mut self, t: &LeafNode<K, V>) -> Self::Output { ... }
}

The Visitor trait allows creating new operations on the radix tree by overriding specific handling methods for each of the node types.

Required Associated Types

type Output

The type of value that the visitor produces.

Required Methods

fn default_output(&self) -> Self::Output

Produce the default value of the Self::Output type.

fn combine_output(&self, o1: Self::Output, o2: Self::Output) -> Self::Output

Combine two instances of the Self::Output type for this Visitor.

Provided Methods

fn visit_node4(&mut self, t: &InnerNode4<K, V>) -> Self::Output

Visit a InnerNode4.

fn visit_node16(&mut self, t: &InnerNode16<K, V>) -> Self::Output

Visit a InnerNode16.

fn visit_node48(&mut self, t: &InnerNode48<K, V>) -> Self::Output

Visit a InnerNode48.

fn visit_node256(&mut self, t: &InnerNode256<K, V>) -> Self::Output

Visit a InnerNode256.

fn visit_leaf(&mut self, t: &LeafNode<K, V>) -> Self::Output

Visit a LeafNode.